Crate atsamd_hal

Re-exports

pub extern crate embedded_hal as hal;

pub use paste;

pub use atsamd51n as target_device;

pub use target_device as pac;

pub use crate::thumbv7em::*;

Modules

common

delay
Delays

gpio
GPIO

pad

prelude
Import the prelude to gain convenient access to helper traits

rtc
Real-time clock/counter

samd51

sercom
Configure the SERCOM peripherals

sleeping_delay
Delays with WFI sleep while we wait using a timer

spi_common

thumbv7em

time
Time units

timer_params
helper struct to calculate divider & cycles settings for timers.

timer_traits

typelevel
Module supporting type-level programming

Macros

bsp_pins
Helper macro to give meaningful names to GPIO pins

dbgprint

define_pins
This macro is a helper for defining a Pins type in a board support crate. This type is used to provide more meaningful aliases for the various GPIO pins for a given board.