Module atsamd_hal::adc

Analogue-to-Digital Conversion

Structs

Adc

Adc encapsulates the device ADC

FreeRunning

InterruptAdc

An ADC where results are accessible via interrupt servicing.

SingleConversion

Enums

Prescaler

Clock frequency relative to the system clock Prescaler Configuration

Reference

Reference voltage (or its source) Reference Selection

Resolution

Reading resolution in bits Conversion Result Resolution

SampleRate

Samples per reading Number of Samples to be Collected

Traits

ConversionMode

Describes how an interrupt-driven ADC should finalize the peripheral upon the completion of a conversion.