atsamd_hal::adc
Struct Adc

pub struct Adc<ADC> { /* private fields */ }

Adc encapsulates the device ADC

Implementations

impl Adc<Adc0>

pub fn adc0( adc: Adc0, mclk: &mut Mclk, clocks: &mut GenericClockController, gclk: Genselect, ) -> Self

pub fn samples(&mut self, samples: SampleRate)

Set the sample rate

pub fn reference(&mut self, reference: Reference)

Set the voltage reference

pub fn prescaler(&mut self, prescaler: Prescaler)

Set the prescaler for adjusting the clock relative to the system clock

pub fn resolution(&mut self, resolution: Resolution)

Set the input resolution

impl Adc<Adc1>

pub fn adc1( adc: Adc1, mclk: &mut Mclk, clocks: &mut GenericClockController, gclk: Genselect, ) -> Self

pub fn samples(&mut self, samples: SampleRate)

Set the sample rate

pub fn reference(&mut self, reference: Reference)

Set the voltage reference

pub fn prescaler(&mut self, prescaler: Prescaler)

Set the prescaler for adjusting the clock relative to the system clock

pub fn resolution(&mut self, resolution: Resolution)

Set the input resolution