atsamd_hal::async_hal::interrupts
Trait Binding

pub unsafe trait Binding<I: InterruptSource, H: Handler<I>> { }

Compile-time assertion that an interrupt has been bound to a handler.

For the vast majority of cases, you should use the bind_interrupts! macro instead of writing unsafe impls of this trait.

Safety

By implementing this trait, you are asserting that you have arranged for H::on_interrupt() to be called every time the I interrupt fires.

This allows drivers to check bindings at compile-time.