Enum atsamd_hal::clock::v2::apb::DynApbId

#[repr(u8)]
pub enum DynApbId {

    Pac,
    Pm,
    Mclk,
    RstC,
    OscCtrl,
    Osc32kCtrl,
    SupC,
    Gclk,
    Wdt,
    Rtc,
    Eic,
    FreqM,
    Sercom0,
    Sercom1,
    Tc0,
    Tc1,
    Usb,
    Dsu,
    NvmCtrl,
    Port,
    EvSys,
    Sercom2,
    Sercom3,
    Tcc0,
    Tcc1,
    Tc2,
    Tc3,
    RamEcc,
    Tcc2,
    PDec,
    Ac,
    Aes,
    Trng,
    Icm,
    Qspi,
    Ccl,
    Sercom4,
    Sercom5,
    Adc0,
    Adc1,
    Dac,
    Pcc,

}

Value-level enum identifying a single APB clock

Each variant of this enum corresponds to a specific bit in one of the four APBXMASK registers and identifies one of many possible APB clocks, which can vary by chip.

DynApbId is the value-level equivalent of ApbId.

Variants

Pac

Pm

Mclk

RstC

OscCtrl

Osc32kCtrl

SupC

Gclk

Wdt

Rtc

Eic

FreqM

Sercom0

Sercom1

Tc0

Tc1

Usb

Dsu

NvmCtrl

Port

EvSys

Sercom2

Sercom3

Tcc0

Tcc1

Tc2

Tc3

RamEcc

Tcc2

PDec

Ac

Aes

Trng

Icm

Qspi

Ccl

Sercom4

Sercom5

Adc0

Adc1

Dac

Pcc