atsamd_hal::clock::v2::apb
Trait ApbId

pub trait ApbId: Sealed {
    const DYN: DynApbId;
}

Type-level enum identifying one of the possible APB clocks

The types implementing this trait are type-level variants of ApbId, and they identify one of the many possible APB clocks, which can vary by chip. Each type corresponds to a specific bit in one of the four APBXMASK registers.

ApbId is the type-level equivalent of DynApbId. See the documentation on type-level programming and specifically type-level enums for more details.

Required Associated Constants

const DYN: DynApbId

Corresponding variant of DynApbId

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.