Trait atsamd_hal::clock::v2::dfll::Reference

pub trait Reference {
    const DYN: DynReference;
}

Type-level enum identifying one of two possible Dfll reference clocks

When the Dfll is in closed-loop mode, it requires a reference clock input. The types implementing this trait, i.e. FromPclk and FromUsb, are type-level variants of Reference, and they identify one of the two possible reference clocks.

Reference is the type-level equivalent of DynReference. See the documentation on type-level programming and specifically type-level enums for more details.

Required Associated Constants

const DYN: DynReference

Corresponding variant of DynReference