Type Definition atsamd_hal::clock::v2::dfll::EnabledDfll

pub type EnabledDfll<I = NoneT, N = U0> = Enabled<Dfll<I>, N>;

An Enabled Dfll

As described in the clock module documentation, the Enabled wrapper implements compile-time clock tree safety by tracking the number of consumer clocks and restricting access to the underlying Dfll to prevent modification while in use.

As with Enabled, the default value for N is U0; if left unspecified, the counter is assumed to be zero.

Implementations

impl<M: Mode> EnabledDfll<M>

pub fn disable(self) -> Dfll<M>

Disable the Dfll

impl<M, N> EnabledDfll<M, N>where
M: Mode,
N: Default,

pub fn into_mode<T, F>(self, mode: T, f: F) -> (EnabledDfll<T, N>, M)where
T: Mode,
F: FnOnce(&mut Dfll<T>),

Change the Dfll Mode while it remains enabled

Take ownership of an EnabledDfll and convert it to use a new Mode. This requires an instance of the new Mode type and returns an instance of the old Mode type. Users can also supply a closure to alter the Dfll settings before they are applied. The closure takes &mut Dfll<T> as its input, so it can only modify those settings with a set_ method.

See the dfll module documentation for more details on why and how this function would be used.