Trait atsamd_hal::clock::v2::dpll::DpllSourceId

pub trait DpllSourceId {
    const DYN: DynDpllSourceId;
}

Type-level enum of possible clock Sources for a Dpll

The types implementing this trait are type-level variants of DpllSourceId, and they identify one of four possible clock Sources for a given Dpll. All implementers of this trait are Id types, which are described in more detail in the clock module documentation.

DpllSourceId is the type-level equivalent of DynDpllSourceId. See the documentation on type-level programming and specifically type-level enums for more details.

Required Associated Constants

const DYN: DynDpllSourceId

Corresponding variant of DynDpllSourceId