atsamd_hal::clock::v2::gclk

Struct GclkOut

pub struct GclkOut<I: GclkIo> { /* private fields */ }

A GPIO Pin configured as a Gclk output

The existence of this struct serves as proof that the corresponding Gclk is Enabled and that it has been output to PinId I.

See the module-level documentation for an example of creating a GclkOut from an EnabledGclk.

Implementations

impl<G, I> GclkOut<I>
where G: GclkId, I: GclkIo<GclkId = G>,

pub fn freq(&self) -> Hertz

Return the frequency of the corresponding Gclk