Trait atsamd_hal::clock::v2::gclk::Gclk0Io

pub trait Gclk0Iowhere
    Self: Sized,
    Self: GclkIo<GclkId = Gclk0Id>,
    Self: GclkSourceId<Resource = Pin<Self, AlternateM>>,{ }

Set of PinIds whose implementations of GclkIo map to Gclk0Id

This is effectively a trait alias for PinIds that implement GclkIo with a GclkId associated type of Gclk0Id, i.e. GclkIo<GclkId = Gclk0Id>. The trait is useful to simply some function signatures and to help type inference in a few cases.