Trait atsamd_hal::clock::v2::gclk::GclkIo

pub trait GclkIo: PinId {
    type GclkId: GclkId;
}

Trait mapping each PinId to its corresponding GclkId when used as a Gclk input or output

If a given PinId can be used as a Gclk input or output, it can only be used with one specific GclkId. This trait provides a mapping from such a PinId to its corresponding GclkId.

Required Associated Types

type GclkId: GclkId

Corresponding GclkId for this PinId