atsamd_hal::clock::v2::gclk

Trait GclkSourceId

pub trait GclkSourceId {
    const DYN: DynGclkSourceId;
}

Type-level enum of possible clock Sources for a Gclk

The types implementing this trait are type-level variants of GclkSourceId, and they identify one of nine possible clock Sources for a given Gclk. All implementers of this trait are Id types, which are described in more detail in the clock module documentation.

GclkSourceId is the type-level equivalent of DynGclkSourceId. See the documentation on type-level programming and specifically type-level enums for more details.

Required Associated Constants

const DYN: DynGclkSourceId

Corresponding variant of DynGclkSourceId

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.