Trait atsamd_hal::clock::v2::gclk::NotGclkIo

pub trait NotGclkIo: GclkSourceId<Resource = ()> { }

Type-level enum of GclkSourceId types that are not a GclkIo

The datasheet notes that a Gclk can use a GPIO Pin as either input or output, but not both. Stated differently, you cannot create a GclkOut from a Gclk where the GclkSourceId is a PinId.

This trait acts as a type-level enum narrowing GclkSourceId to exclude any types which implement GclkIo.