Type Definition atsamd_hal::clock::v2::gclk::EnabledGclk

pub type EnabledGclk<G, I, N = U0> = Enabled<Gclk<G, I>, N>;

An Enabled Gclk

As described in the clock module documentation, the Enabled wrapper implements compile-time clock tree safety by tracking the number of clocks consuming this Gclk and restricts access to the underlying Gclk to prevent misuse.

As with Enabled, the default value for N is U0; if left unspecified, the counter is assumed to be zero.

Implementations

impl<G, I> EnabledGclk<G, I>where
G: GclkId,
I: GclkSourceId,

pub fn disable(self) -> Gclk<G, I>

Disable the Gclk

This method is only implemented for N = U0, which means the clock can only be disabled when no other clocks consume this Gclk.

impl<G, S, N> EnabledGclk<G, S, N>where
G: GclkId,
S: NotGclkIo,

pub fn enable_gclk_out<P>(
self,
pin: P
) -> (EnabledGclk<G, S, N::Inc>, GclkOut<P::Id>)where
N: Increment,
P: AnyPin,
P::Id: GclkIo<GclkId = G>,

Create and enable a GclkOut

Enabling GclkIo output will Increment the EnabledGclk counter, which will prevent it from being disabled while the GclkOut exists.

Note that a given Gclk can only use GclkIo for input or output, but not both simultaneously. The NotGclkIo trait exists to enforce this requirement.

Finally, when a GclkOut is disabled, but the Pin is still in AlternateM mode, it takes the “output off value” of the Gclk. See the Gclk::output_off_value documentation for more details.

pub fn disable_gclk_out<I>(
self,
gclk_out: GclkOut<I>
) -> (EnabledGclk<G, S, N::Dec>, Pin<I, AlternateM>)where
N: Decrement,
I: GclkIo<GclkId = G>,

Disable a GclkOut and free its Pin

Disabling GclkIo output will Decrement the EnabledGclk counter. When a GclkOut is disabled, but the Pin is still in AlternateM mode, it takes the “output off value” of the Gclk. See the Gclk::output_off_value documentation for more details.