Type Definition atsamd_hal::clock::v2::osculp32k::EnabledOscUlp1k

pub type EnabledOscUlp1k<N = U0> = Enabled<OscUlp1k, N>;

The Enabled OscUlp1k clock

As described in the clock module documentation, the Enabled wrapper implements compile-time clock tree safety by tracking the number of clocks consuming the OscUlp1k clock and restricts access to the underlying type to prevent misuse.

As with Enabled, the default value for N is U0; if left unspecified, the counter is assumed to be zero.

Implementations

impl EnabledOscUlp1k

pub fn disable<N: Decrement>(
self,
base: EnabledOscUlp32kBase<N>
) -> (OscUlp1kToken, EnabledOscUlp32kBase<N::Dec>)

Disable 1 kHz output from the OscUlp32kBase clock

This will Decrement the EnabledOscUlp32kBase counter.