Type Definition atsamd_hal::clock::v2::osculp32k::EnabledOscUlp32kBase

pub type EnabledOscUlp32kBase<N = U0> = Enabled<OscUlp32kBase, N>;

The Enabled OscUlp32kBase clock

As described in the clock module documentation, the Enabled wrapper implements compile-time clock tree safety by tracking the number of clocks consuming the OscUlp32kBase clock and restricts access to the underlying type to prevent misuse.

NOTE: The OscUlp32kBase clock is internal and can never be disabled, so we do not provide a disable method.

As with Enabled, the default value for N is U0; if left unspecified, the counter is assumed to be zero.

Implementations

impl<N> EnabledOscUlp32kBase<N>

pub fn set_calibration(&mut self, calib: u8)

Override the factory-default calibration value

pub fn write_lock(self)

Freeze the OSCULP32K configuration until power-on reset

This function sets the write-lock bit, which freezes the OSCULP32K configuration at the hardware level until power-on reset. At the API level, it also consumes and drops the OscUlp32kBase clock, which prevents any further modifications.