Module atsamd_hal::clock::v2::pclk::ids

Module containing only the types implementing PclkId

Because there are so many types that implement PclkId, it is helpful to have them defined in a separate module, so that you can import all of them using a wildcard (*) without importing anything else, i.e.

use atsamd_hal::clock::v2::pclk::ids::*;

Re-exports

pub use crate::sercom::Sercom0;

pub use crate::sercom::Sercom1;

pub use crate::sercom::Sercom2;

pub use crate::sercom::Sercom3;

pub use crate::sercom::Sercom4;

pub use crate::sercom::Sercom5;

pub use crate::sercom::Sercom6;

pub use crate::sercom::Sercom7;

Enums

Ac

Marker type representing the corresponding peripheral

Adc0

Marker type representing the corresponding peripheral

Adc1

Marker type representing the corresponding peripheral

CM4Trace

Marker type representing the corresponding peripheral

Ccl

Marker type representing the corresponding peripheral

Dac

Marker type representing the corresponding peripheral

DfllId

Id type representing the identity of the DFLL clock

Dpll0Id

Type-level variant of DpllId representing the identity of DPLL0

Dpll1Id

Type-level variant of DpllId representing the identity of DPLL1

Eic

Marker type representing the corresponding peripheral

EvSys0

Marker type representing the corresponding peripheral

EvSys1

Marker type representing the corresponding peripheral

EvSys2

Marker type representing the corresponding peripheral

EvSys3

Marker type representing the corresponding peripheral

EvSys4

Marker type representing the corresponding peripheral

EvSys5

Marker type representing the corresponding peripheral

EvSys6

Marker type representing the corresponding peripheral

EvSys7

Marker type representing the corresponding peripheral

EvSys8

Marker type representing the corresponding peripheral

EvSys9

Marker type representing the corresponding peripheral

EvSys10

Marker type representing the corresponding peripheral

EvSys11

Marker type representing the corresponding peripheral

FreqMMeasure

Marker type representing the corresponding peripheral

FreqMReference

Marker type representing the corresponding peripheral

I2S0

Marker type representing the corresponding peripheral

I2S1

Marker type representing the corresponding peripheral

PDec

Marker type representing the corresponding peripheral

Sdhc0

Marker type representing the corresponding peripheral

Sdhc1

Marker type representing the corresponding peripheral

SlowClk

Marker type representing the corresponding peripheral

Tc0Tc1

Marker type representing the corresponding peripheral

Tc2Tc3

Marker type representing the corresponding peripheral

Tc4Tc5

Marker type representing the corresponding peripheral

Tc6Tc7

Marker type representing the corresponding peripheral

Tcc0Tcc1

Marker type representing the corresponding peripheral

Tcc2Tcc3

Marker type representing the corresponding peripheral

Tcc4

Marker type representing the corresponding peripheral

Usb

Marker type representing the corresponding peripheral