Struct atsamd_hal::clock::v2::pclk::Pclk

pub struct Pclk<P, I>where
    P: PclkId,
    I: PclkSourceId,{ /* private fields */ }

Peripheral channel clock for a given peripheral

Peripheral channel clocks connect generic clock generators (Gclks) to various peripherals. Pclks usually act as leaves in the clock tree, except when they feed the DFLL and DPLL peripherals.

The type parameter P is a PclkId that determines which of the 48 peripherals this Pclk feeds. The type parameter I represents the Id type for the EnabledGclk acting as the Pclk’s Source. It must be one of the valid PclkSourceIds, which is simply a trait alias for GclkId. See the clock module documentation for more detail on Id types.

Pclks cannot act as general purpose clock Sources; rather, they map 1:1 with corresponding peripherals. Thus, enabled Pclks do not need a compile-time counter of consumer clocks, so they are not wrapped with Enabled. Enabled Pclks are created directly from PclkTokens with Pclk::enable.

See the module-level documentation for an example.

Implementations

impl<P, I> Pclk<P, I>where
P: PclkId,
I: PclkSourceId,

pub fn enable<S>(token: PclkToken<P>, gclk: S) -> (Self, S::Inc)where
S: Source<Id = I> + Increment,

Create and enable a Pclk

Creating a Pclk immediately enables the corresponding peripheral channel clock. It also Increments the Source’s Enabled counter.

Note that the Source will always be an EnabledGclk.

pub fn disable<S>(self, gclk: S) -> (PclkToken<P>, S::Dec)where
S: Source<Id = I> + Decrement,

Disable and destroy a Pclk

Consume the Pclk, release the PclkToken, and Decrement the EnabledGclk’s counter

pub fn freq(&self) -> Hertz

Return the Pclk frequency