Trait atsamd_hal::clock::v2::pclk::PclkSourceId

pub trait PclkSourceId: GclkId { }

Type-level enum of possible clock Sources for a Pclk

The types implementing this trait are type-level variants of PclkSourceId, and they identify the Gclk acting as a clock Source for a given Pclk. Accordingly, all implementers of this trait are GclkId types, and this trait is simply a trait alias for GclkId. Id types in general are described in more detail in the clock module documentation.

PclkSourceId is the type-level equivalent of DynPclkSourceId. See the documentation on type-level programming and specifically type-level enums for more details.