atsamd_hal::clock::v2::pclk

Type Alias DynPclkSourceId

pub type DynPclkSourceId = DynGclkId;

Value-level enum of possible clock sources for a Pclk

The variants of this enum identify the Gclk used as a clock source for a given Pclk. Because the variants are identical to DynGclkId, we simply define it as a type alias.

DynPclkSourceId is the value-level equivalent of PclkSourceId.

Aliased Type

enum DynPclkSourceId {
    Gclk0,
    Gclk1,
    Gclk2,
    Gclk3,
    Gclk4,
    Gclk5,
    Gclk6,
    Gclk7,
    Gclk8,
    Gclk9,
    Gclk10,
    Gclk11,
}

Variants

Gclk0

Gclk1

Gclk2

Gclk3

Gclk4

Gclk5

Gclk6

Gclk7

Gclk8

Gclk9

Gclk10

Gclk11