Enum atsamd_hal::clock::v2::rtcosc::DynRtcSourceId

pub enum DynRtcSourceId {
    OscUlp1k,
    OscUlp32k,
    Xosc1k,
    Xosc32k,
}

Value-level enum of possible clock sources for the RtcOsc

The variants of this enum identify one of four possible clock sources for the RtcOsc.

DynRtcSourceId is the value-level equivalent of RtcSourceId.

Variants

OscUlp1k

The RTC is sourced from the OscUlp1k

OscUlp32k

The RTC is sourced from the OscUlp32k

Xosc1k

The RTC is sourced from the Xosc1k

Xosc32k

The RTC is sourced from the Xosc32k