atsamd_hal::clock::v2::xosc32k

Type Alias EnabledXosc32k

pub type EnabledXosc32k<N = U0> = Enabled<Xosc32k, N>;

The Enabled Xosc32k clock

As described in the clock module documentation, the Enabled wrapper implements compile-time clock tree safety by tracking the number of clocks consuming the Xosc32k clock and restricts access to the underlying type to prevent misuse.

As with Enabled, the default value for N is U0; if left unspecified, the counter is assumed to be zero.

Aliased Type

struct EnabledXosc32k<N = U0>(/* private fields */);

Implementations

impl EnabledXosc32k

pub fn disable<M, N>( self, base: EnabledXosc32kBase<M, N>, ) -> (Xosc32kToken, EnabledXosc32kBase<M, N::Dec>)
where M: Mode, N: Decrement,

Disable 1 kHz output from the Xosc32kBase clock

This will Decrement the EnabledXosc32kBase counter.