Type Definition atsamd_hal::clock::v2::xosc32k::EnabledXosc32kBase

pub type EnabledXosc32kBase<M, N = U0> = Enabled<Xosc32kBase<M>, N>;

The Enabled Xosc32kBase clock

As described in the clock module documentation, the Enabled wrapper implements compile-time clock tree safety by tracking the number of clocks consuming the Xosc32kBase clock and restricts access to the underlying type to prevent misuse.

As with Enabled, the default value for N is U0; if left unspecified, the counter is assumed to be zero.

Implementations

impl<M: Mode> EnabledXosc32kBase<M>

pub fn disable(self) -> Xosc32kBase<M>

Disable the Xosc32kBase clock

This method is only implemented for N = U0, which means the clock can only be disabled when no other clocks consume this Xosc32kBase clock.

impl<M: Mode, N> EnabledXosc32kBase<M, N>

pub fn is_ready(&self) -> bool

Check whether the XOSC32K is stable and ready to be used as a clock Source

pub fn write_lock(self)

Freeze the XOSC32K configuration until power-on reset

This function sets the write-lock bit, which freezes the XOSC32K configuration at the hardware level until power-on reset. At the API level, it also consumes and drops the Xosc32kBase clock, which prevents any further modifications.