Module atsamd_hal::common

Modules

adc

calibration
NVM Software Calibration Area Mapping

clock
Configuring the system clock sources. You will typically need to create an instance of GenericClockController before you can set up most of the peripherals on the atsamd51 device. The other types in this module are used to enforce at compile time that the peripherals have been correctly configured.

delay
Delays

eic

gpio
GPIO

pad

prelude
Import the prelude to gain convenient access to helper traits

pwm

qspi

rtc
Real-time clock/counter

sercom
Configure the SERCOM peripherals

sleeping_delay
Delays with WFI sleep while we wait using a timer

spi_common

time
Time units

timer
Working with timer counter hardware

timer_params
helper struct to calculate divider & cycles settings for timers.

timer_traits

trng

usb
USB Device support

watchdog

Enums

ResetCause
ResetCause represents the reason the MCU was reset.

Functions

reset_cause
Returns the cause of the last reset.

serial_number
Returns the serial number of the chip as an array of bytes. The serial number is only guaranteed to be unique if all 16 bytes are used.

split_serial_number
Returns the serial number of the chip as 4 32-bit integers. The serial number is only guaranteed to be unique if all 128 bits are used.