Struct atsamd_hal::common::gpio::v2::dynpin::DynPin

pub struct DynPin { /* fields omitted */ }

A value-level pin, parameterized by DynPinId and DynPinMode

This type acts as a type-erased version of super::Pin. Every pin is represented by the same type, and pins are tracked and distinguished at run-time.

Implementations

impl DynPin

pub fn id(&self) -> DynPinId

Return a copy of the pin ID

pub fn mode(&self) -> DynPinMode

Return a copy of the pin mode

pub fn into_mode(&mut self, mode: DynPinMode)

Convert the pin to the requested DynPinMode

pub fn into_floating_disabled(&mut self)

Disable the pin and set it to float

pub fn into_pull_down_disabled(&mut self)

Disable the pin and set it to pull down

pub fn into_pull_up_disabled(&mut self)

Disable the pin and set it to pull up

pub fn into_floating_input(&mut self)

Configure the pin to operate as a floating input

pub fn into_pull_down_input(&mut self)

Configure the pin to operate as a pulled down input

pub fn into_pull_up_input(&mut self)

Configure the pin to operate as a pulled up input

pub fn into_push_pull_output(&mut self)

Configure the pin to operate as a push-pull output

pub fn into_readable_output(&mut self)

Configure the pin to operate as a readable push pull output

pub fn into_alternate(&mut self, config: DynAlternate)

Configure the pin to operate as the corresponding peripheral function.

The config parameter indicates the desired peripheral function.

pub fn get_drive_strength(&self)

Read the current drive strength of the pin.

The drive strength is reset to normal on every change in pin mode.

pub fn set_drive_strength(&mut self, stronger: bool)

Set the drive strength for the pin.

The drive strength is reset to normal on every change in pin mode.