Struct atsamd_hal::common::gpio::v2::pin::Pin

pub struct Pin<I, M> where
    I: PinId,
    M: PinMode,  { /* fields omitted */ }

A type-level GPIO pin, parameterized by PinId and PinMode types

Implementations

impl<I, M> Pin<I, M> where
I: PinId,
M: PinMode,

pub fn as_any<P>(self) -> P where
P: AnyPin<Id = I, Mode = M>,

Convert a Pin to a type that implements AnyPin

Even though there is a one-to-one mapping between Pin<I, M> and AnyPin<Id = I, Mode = M>, the compiler doesn't know that. This method provides a way to convert from a Pin to an AnyPin. See the AnyPin trait for more details.

pub fn as_any_ref<P>(&self) -> &P where
P: AnyPin<Id = I, Mode = M>,

Implement as_any for references

pub fn as_any_mut<P>(&mut self) -> &mut P where
P: AnyPin<Id = I, Mode = M>,

Implement as_any for mutable references

pub fn into_mode<N: PinMode>(self) -> Pin<I, N>

Convert the pin to the requested PinMode

pub fn into_floating_disabled(self) -> Pin<I, FloatingDisabled>

Disable the pin and set it to float

pub fn into_pull_down_disabled(self) -> Pin<I, PullDownDisabled>

Disable the pin and set it to pull down

pub fn into_pull_up_disabled(self) -> Pin<I, PullUpDisabled>

Disable the pin and set it to pull up

pub fn into_floating_input(self) -> Pin<I, FloatingInput>

Configure the pin to operate as a floating input

pub fn into_pull_down_input(self) -> Pin<I, PullDownInput>

Configure the pin to operate as a pulled down input

pub fn into_pull_up_input(self) -> Pin<I, PullUpInput>

Configure the pin to operate as a pulled up input

pub fn into_push_pull_output(self) -> Pin<I, PushPullOutput>

Configure the pin to operate as a push-pull output

pub fn into_readable_output(self) -> Pin<I, ReadableOutput>

Configure the pin to operate as a readable push pull output

pub fn into_alternate<C: AlternateConfig>(self) -> Pin<I, Alternate<C>>

Configure the pin to operate as the corresponding peripheral function.

The type C indicates the desired peripheral function.

pub fn get_drive_strength(&self)

Read the current drive strength of the pin.

The drive strength is reset to normal on every change in pin mode.

pub fn set_drive_strength(&mut self, stronger: bool)

Set the drive strength for the pin.

The drive strength is reset to normal on every change in pin mode.