Trait atsamd_hal::common::gpio::v2::pin::SomePin

pub trait SomePin: OptionalPin + AnyPin { }

Meta-type representing a valid Pin.

When used as a bound, this trait allows you to exclude NoneT and limit the type to valid Pins. AnyPin is a super trait to SomePin, so all of its functionality is still available.