Trait atsamd_hal::common::sercom::v1::pads::PadPin

pub trait PadPin<T>: Sealed {
    fn into_pad(self, port: &mut Port) -> T;
}

The PadPin trait makes it more ergonomic to convert a pin into a Sercom pad. You should not implement this trait for yourself; only the implementations in the sercom module make sense.

Required methods

fn into_pad(self, port: &mut Port) -> T

Loading content...