Struct atsamd_hal::common::sercom::v1::spi::SPIMaster0Padout

pub struct SPIMaster0Padout<MISO, MOSI, SCK> { /* fields omitted */ }

A pad mapping configuration for the SERCOM in SPI master mode.

This type can only be constructed using the From implementations in this module, which are restricted to valid configurations.

Defines which sercom pad is mapped to which SPI function.