Struct atsamd_hal::common::sercom::v1::SPIMaster7

pub struct SPIMaster7<MISO, MOSI, SCK> { /* fields omitted */ }

SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

This type is generic over any valid pad mapping where there is a defined "data in pin out data out pin out" implementation.

Implementations

impl<MISO, MOSI, SCK> SPIMaster7<MISO, MOSI, SCK>

pub fn new<F: Into<Hertz>, T: Into<SPIMaster7Padout<MISO, MOSI, SCK>>>(
clock: &Sercom7CoreClock,
freq: F,
mode: Mode,
sercom: SERCOM7,
mclk: &mut MCLK,
padout: T
) -> Self where
SPIMaster7Padout<MISO, MOSI, SCK>: DipoDopo,

Power on and configure SERCOMX to work as an SPI Master operating with the specified frequency and SPI Mode. The pinout specifies which pins are bound to the MISO, MOSI, SCK functions.

pub fn set_baud<F: Into<Hertz>>(&mut self, freq: F, clock: &Sercom7CoreClock)

Set the baud rate

pub fn free(self) -> (SPIMaster7Padout<MISO, MOSI, SCK>, SERCOM7)

Tear down the SPI instance and yield the constituent pins and SERCOM instance. No explicit de-initialization is performed.