Struct atsamd_hal::common::sercom::v1::UART4

pub struct UART4<RX, TX, RTS, CTS> { /* fields omitted */ }

UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

This type is generic over any valid pad mapping where there is a defined "receive pin out transmit pin out" implementation.

Implementations

impl<RX, TX, RTS, CTS> UART4<RX, TX, RTS, CTS>

pub fn new<F: Into<Hertz>, T: Into<UART4Padout<RX, TX, RTS, CTS>>>(
clock: &Sercom4CoreClock,
freq: F,
sercom: SERCOM4,
mclk: &mut MCLK,
padout: T
) -> Self where
UART4Padout<RX, TX, RTS, CTS>: RxpoTxpo,

pub fn free(self) -> (UART4Padout<RX, TX, RTS, CTS>, SERCOM4)

pub fn intenset<F>(&mut self, f: F) where
F: FnOnce(&mut W),

pub fn intenclr<F>(&mut self, f: F) where
F: FnOnce(&mut W),

pub fn flags(&self) -> R