Module atsamd_hal::common::sercom::v1::uart

Structs

UART0
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART0Padout
A pad mapping configuration for the SERCOM in UART mode.

UART1
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART1Padout
A pad mapping configuration for the SERCOM in UART mode.

Traits

RxpoTxpo
The RxpoTxpo trait defines a way to get the data in and data out pin out values for a given UARTXPadout configuration. You should not implement this trait for yourself; only the implementations in the sercom module make sense.