Trait atsamd_hal::common::sercom::v1::uart::RxpoTxpo

pub trait RxpoTxpo {
    fn rxpo_txpo(&self) -> (u8, u8);
}

The RxpoTxpo trait defines a way to get the data in and data out pin out values for a given UARTXPadout configuration. You should not implement this trait for yourself; only the implementations in the sercom module make sense.

Required methods

fn rxpo_txpo(&self) -> (u8, u8)

Loading content...