Module atsamd_hal::common::sercom::v2

Version 2 of the SERCOM module

This module provides a new API for the SERCOM peripherals. So far, only the pads module has been updated, but it is expected that the uart, spi and i2c modules will receive updates as well.

Re-exports

pub use pads::*;

Modules

pad_map
Implementations of the pads Map trait

pads
Version 2 of the SERCOM pads module