Module atsamd_hal::common::sercom::v2::pad_map

Implementations of the pads Map trait

This module provides implementations of Map on each IoSet for each valid pair of Sercom and PadNum. Each combination of Sercom, PadNum and IoSet maps to a unique PinId and PinMode.