Struct atsamd_hal::common::sercom::v2::pads::Pad

pub struct Pad<S, P, M> where
    S: Sercom,
    P: PadNum,
    M: Map<S, P>,  { /* fields omitted */ }

Represents a SERCOM Pad configured to use a particular pin

Each Pad is parameterized by a Sercom, a PadNum, and a third type that implements Map, which is used to determine the corresponding Pin and its configuration.

For the SAMD51 and SAME5x chips, Map is implemented on instances of IoSet . The SAMD11 and SAMD21 do not have any concept of IOSET, so Map is instead implemented directly on the corresponding PinId.

Each Pad takes ownership of the corresponding Pin for the duration of its lifetime. Pads can be converted to and from Pins using the Into and From traits.

Implementations

impl<S, P, M> Pad<S, P, M> where
S: Sercom,
P: PadNum,
M: Map<S, P>,

pub fn new<O: PinMode>(pin: Pin<M::Id, O>) -> Self where
Pin<M::Id, O>: Into<Pin<M::Id, M::Mode>>,

Create a new SERCOM Pad from a Pin

The specified Map type must map the specified Sercom and PadNum to the given Pin

pub fn free(self) -> Pin<M::Id, M::Mode>

Consume the Pad and release the corresponding Pin

pub fn as_any<T>(self) -> T where
T: AnyPad<Sercom = S, PadNum = P, Map = M>,

Convert a Pad to a type that implements AnyPad

Even though there is a one-to-one mapping between Pad<S, P, M> and AnyPad<Sercom = S, PadNum = P, Map = M>, the compiler doesn't know that. This method provides a way to convert from a Pad to an AnyPad. See the AnyPad trait for more details.