Trait atsamd_hal::common::sercom::v2::pads::Map

pub trait Map<S, P> where
    S: Sercom,
    P: PadNum, {
    type Id: PinId;
    type Mode: PinMode;
}

Type-level function mapping Pads to Pins

This trait acts as a type-level function. It takes two types as arguments, the Sercom and PadNum of a Pad, and returns the PinId and PinMode for the corresponding Pin.

For the SAMD51 and SAME5x series chips, all pins for a given SERCOM must come from the same IOSET. To account for this, we introduce a new IoSet meta-type and implement Map on its instances. For a given Sercom and PadNum, the IoSet uniquely identifies a corresponding PinId and PinMode.

The SAMD11 and SAMD21 series chips are not limited by IOSET. Any combination of valid pins for a given SERCOM is acceptable. Thus, the Map trait is implemented directly on PinIds. Because the same Pin can often be used for two different Pads, the Map trait acts to map a Sercom/PadNum pair to the correct PinMode for the PinId.

Associated Types

type Id: PinId

The PinId for the corresponding pin

type Mode: PinMode

The PinMode for the corresponding pin

Loading content...