Trait atsamd_hal::common::timer::Count16

pub trait Count16 {
    fn count_16(&self) -> &COUNT16;
}

This is a helper trait to make it easier to make most of the TimerCounter impl generic. It doesn't make too much sense to to try to implement this trait outside of this module.

Required methods

fn count_16(&self) -> &COUNT16

Loading content...