atsamd_hal::dmac::dma_controller

Struct DmaController

pub struct DmaController<I = NoneT> { /* private fields */ }

Initialized DMA Controller

Implementations

impl DmaController

pub fn init(dmac: Dmac, _pm: &mut Pm) -> Self

Initialize the DMAC and return a DmaController object useable by Transfer’s. By default, all priority levels are enabled unless subsequently disabled using the level_x_enabled methods.

pub fn free(self, _channels: Channels, _pm: &mut Pm) -> Dmac

Release the DMAC and return the register block.

Note: The Channels struct is consumed by this method. This means that any Channel obtained by split must be moved back into the Channels struct before being able to pass it into free.

impl<T> DmaController<T>

pub fn enable_levels(&mut self, mask: PriorityLevelMask)

Enable multiple priority levels simultaneously

pub fn disable_levels(&mut self, mask: PriorityLevelMask)

Disable multiple priority levels simultaneously

pub fn round_robin_arbitration(&mut self, mask: RoundRobinMask)

Enable round-robin arbitration for multiple priority levels simultaneously

pub fn static_arbitration(&mut self, mask: RoundRobinMask)

Disable round-robin arbitration (ie, enable static priorities) for multiple priority levels simultaneously

pub fn into_future<I>(self, _interrupts: I) -> DmaController<I>
where I: Binding<DMAC, InterruptHandler>,

Use the DmaController in async mode. You are required to provide the struct created by the bind_interrupts macro to prove that the interrupt sources have been correctly configured. This function will automatically enable the relevant NVIC interrupt sources. However, you are required to configure the desired interrupt priorities prior to calling this method. Consult crate::async_hal::interrupts module-level documentation for more information. bind_interrupts.

impl<I> DmaController<I>
where I: Binding<DMAC, InterruptHandler>,

pub fn free(self, _channels: FutureChannels, _pm: &mut Pm) -> Dmac

Release the DMAC and return the register block.

Note: The Channels struct is consumed by this method. This means that any Channel obtained by split must be moved back into the Channels struct before being able to pass it into free.

impl DmaController

pub fn split(&mut self) -> Channels

Split the DMAC into individual channels

impl<I> DmaController<I>
where I: Binding<DMAC, InterruptHandler>,

pub fn split(&mut self) -> FutureChannels

Split the DMAC into individual channels