atsamd_hal::dmac
Module transfer

DMA transfer abstractions

Transfer types

Four basic transfer types are supported:

Incrementing-source to incrementing-destination (normally used for memory-to-memory transfers)

Incrementing-source to fixed-destination (normally used for memory-to-peripheral transfers)

Fixed-source to incrementing-destination (normally used for peripheral-to-memory transfers)

Fixed-source to fixed-destination (normally used for peripheral-to-peripheral transfers)

Beat sizes

A beat is an atomic, uninterruptible transfer size.Three beat sizes are supported:

Byte-per-byte (8 bit beats);

Halfword-per-halfword (16 bit beats);

Word-per-word (32 bit beats);

The correct beat size will automatically be selected in function of the type of the source and destination buffers.

One-shot vs circular transfers

If the transfer is setup as one-shot (circular == false), the transfer will run once and stop. Otherwise, if circular == true, then the transfer will be set as circular, meaning it will restart a new, identical block transfer when the current block transfer is complete. This is particularly useful when combined with a TC/TCC trigger source, for instance to periodically retreive a sample from an ADC and send it to a circular buffer, or send a sample to a DAC.

Starting a transfer

A transfer is started by calling Transfer::begin. You will be required to supply a trigger source and a trigger action.

Waiting for a transfer to complete

A transfer can waited upon by calling wait. This is a blocking method, meaning it will busy-wait until the transfer is completed. When it returns, it will release the source and destination buffers, as well as the DMA channel.

Interrupting (stopping) a transfer

A transfer can be stopped (regardless of whether it has completed or not) by calling stop. This is not a blocking method, meaning it will stop the transfer and immediately return. When it returns, it will release the source and destination buffers, as well as the DMA channel.

Trigger sources

Most peripherals can issue triggers to a DMA channel. A software trigger is also available (see trigger). See ATSAMD21 datasheet, table 19-8 for all available trigger sources.

Trigger actions

Three trigger actions are available:

BLOCK: One trigger required for each block transfer. In the context of this driver, one Transfer is equivalent to one Block transfer.

BEAT: One trigger required for each beat transfer. In the context of this driver, the beat size will depend on the type of buffer used (8, 16 or 32 bits).

TRANSACTION: One trigger required for a full DMA transaction. this is useful for circular transfers in the context of this driver. One trigger will set off the transaction, that will now run uninterrupted until it is stopped.

Structs

BufferPair

Struct holding the source and destination buffers of a Transfer.

Transfer

DMA transfer, owning the resources until the transfer is done and Transfer::wait is called.

Enums

BeatSize

Useable beat sizes for DMA transfers

Traits

AnyBufferPair

Beat

Convert 8, 16 and 32 bit types into BeatSize

Buffer

Buffer useable by the DMAC.

Type Aliases

BufferPairBeat

BufferPairDst

BufferPairSrc

SpecificBufferPair