atsamd_hal::eic
Struct Eic

pub struct Eic<I = NoneT> { /* private fields */ }

External Interrupt Controller.

Use split to split the struct into individual channels, which can then be used to create ExtInts, by calling Channel::with_pin.

Implementations

impl Eic

pub fn into_future<I>(self, _irq: I) -> Eic<EicFuture>
where I: Binding<EicInterrupt, InterruptHandler>,

Turn an EIC pin into a pin usable as a Future. The correct interrupt source is needed.

impl Eic

pub fn new(pm: &mut Pm, _clock: EicClock, eic: Eic) -> Self

Create a new Eic and initialize it.

pub fn free(self, _channels: Channels) -> Eic

Release the EIC and return the register block.

Note: The Channels struct is consumed by this method. This means that any Channel obtained by split must be moved back into the Channels struct before being able to pass it into free.

impl Eic<EicFuture>

pub fn free(self, _channels: FutureChannels) -> Eic

Release the EIC and return the register block.

Note: The Channels struct is consumed by this method. This means that any Channel obtained by split must be moved back into the Channels struct before being able to pass it into free.

impl Eic

pub fn split(self) -> Channels

Split the EIC into individual channels.

impl Eic<EicFuture>

pub fn split(self) -> FutureChannels

Split the EIC into individual channels