atsamd_hal::eic
Struct ExtInt

pub struct ExtInt<P, Id, F = NoneT>
where
    P: EicPin,
    Id: ChId,
{ /* private fields */ }

A numbered external interrupt, which can be used to sense state changes on its pin.

Implementations

impl<P, Id> ExtInt<P, Id, EicFuture>
where P: EicPin, Id: ChId, Self: InputPin<Error = Infallible>,

pub async fn wait(&mut self, sense: Sense)

impl<P, Id, F> ExtInt<P, Id, F>
where P: EicPin, Id: ChId,

pub fn enable_event(&mut self)

Configure the eic with options for this external interrupt

pub fn enable_interrupt(&mut self)

pub fn enable_interrupt_wake(&mut self)

pub fn disable_interrupt(&mut self)

pub fn is_interrupt(&mut self) -> bool

pub fn clear_interrupt(&mut self)

pub fn sense(&mut self, sense: Sense)

pub fn filter(&mut self, filter: bool)

impl<P, Id, F> ExtInt<P, Id, F>
where P: EicPin, Id: ChId,

pub fn free(self) -> (Pin<P::Id, P::Mode>, Channel<Id, F>)

Release the underlying resources: Pin and Channel.