Struct atsamd_hal::gpio::v1::Pin

pub struct Pin<I, M> where
    I: PinId,
    M: PinMode,  { /* fields omitted */ }

Represents a GPIO pin with a corresponding PinId and PinMode

The v2::Pin type provides many of the same inherent functions, but it does so without requiring the PORT as an argument, breaking backwards compatibility.

v1 Pin type is a newtype wrapper for v2::Pins. To aid in compatibility, the v1 Pin types also implement AnyPin. From & Into conversions are provided between the two pin types.

Implementations

impl<I, M> Pin<I, M> where
I: PinId,
M: PinMode,

pub fn into_floating_input(self, port: &mut Port) -> Pin<I, Input<Floating>>

Configures the pin to operate as a floating input

pub fn into_pull_down_input(self, port: &mut Port) -> Pin<I, Input<PullDown>>

Configures the pin to operate as a pulled down input pin

pub fn into_pull_up_input(self, port: &mut Port) -> Pin<I, Input<PullUp>>

Configures the pin to operate as a pulled up input pin

pub fn into_floating_interrupt(
self,
port: &mut Port
) -> Pin<I, Interrupt<Floating>>

Configures the pin to operate as a floating interrupt

pub fn into_pull_down_interrupt(
self,
port: &mut Port
) -> Pin<I, Interrupt<PullDown>>

Configures the pin to operate as a pulled down interrupt pin

pub fn into_pull_up_interrupt(
self,
port: &mut Port
) -> Pin<I, Interrupt<PullUp>>

Configures the pin to operate as a pulled up interrupt pin

pub fn into_open_drain_output(
self,
port: &mut Port
) -> Pin<I, Output<OpenDrain>>

Configures the pin to operate as an open drain output

pub fn into_readable_open_drain_output(
self,
port: &mut Port
) -> Pin<I, Output<ReadableOpenDrain>>

Configures the pin to operate as an open drain output which can be read

pub fn into_push_pull_output(self, port: &mut Port) -> Pin<I, Output<PushPull>>

Configures the pin to operate as a push-pull output

pub fn into_function_b(self, port: &mut Port) -> Pin<I, PfB>

Configures the pin to operate with a peripheral

pub fn into_function_c(self, port: &mut Port) -> Pin<I, PfC>

Configures the pin to operate with a peripheral

pub fn into_function_d(self, port: &mut Port) -> Pin<I, PfD>

Configures the pin to operate with a peripheral

pub fn into_function_e(self, port: &mut Port) -> Pin<I, PfE>

Configures the pin to operate with a peripheral

pub fn into_function_f(self, port: &mut Port) -> Pin<I, PfF>

Configures the pin to operate with a peripheral

pub fn into_function_g(self, port: &mut Port) -> Pin<I, PfG>

Configures the pin to operate with a peripheral

impl<I: PinId> Pin<I, Output<OpenDrain>>

pub fn internal_pull_up(&mut self, port: &mut Port, on: bool)

Control state of the internal pull up

This function shouldn’t exist. It is not possible to enable a pull-up resistor in an output mode.

impl<I, M> Pin<I, Output<M>> where
I: PinId,
M: OutputConfig,

pub fn toggle(&mut self)

Toggle the logic level of the pin; if it is currently high, set it low and vice versa.