Trait atsamd_hal::gpio::v2::pin::PinMode

pub trait PinMode: Sealed {
    const DYN: DynPinMode;
}

Type-level enum representing pin modes

The valid options are Disabled, Input, Output and Alternate. See the type-level enum documentation for more details on the pattern.

Associated Constants

const DYN: DynPinMode

Corresponding DynPinMode

Loading content...