Struct atsamd_hal::icm::Region

pub struct Region<I: RegionNum> { /* private fields */ }

Region provides access to region-specific settings like interrupts and status

Implementations

impl<I: RegionNum> Region<I>

pub fn enable_monitoring(&self)

Enable this memory monitor region

pub fn get_monitoring_status(&self) -> bool

Returns true if the region monitoring is active and a check value has been calculated and written to the hash area

pub fn get_monitoring_raw_status(&self) -> bool

Returns true if the RMEN register has been set to one for the region

pub fn disable_monitoring(&self)

Disable the memory monitor region

pub fn trigger_rehash(&self)

Trigger recalculation of the memory monitor region

pub fn set_rsu_int(&self)

Set Region Status Updated interrupt enable

pub fn set_rec_int(&self)

Set Region End bit Condition Detected interrupt enable

pub fn set_rwc_int(&self)

Set Region Wrap Condition detected interrupt enable

pub fn set_rbe_int(&self)

Set Region Bus Error interrupt enable

pub fn set_rdm_int(&self)

Set Region Digest Mismatch interrupt enable

pub fn set_rhc_int(&self)

Set Region Hash Completed interrupt enable

pub fn disable_rsu_int(&self)

Disable Region Status Updated interrupt enable

pub fn disable_rec_int(&self)

Disable Region End bit Condition Detected interrupt enable

pub fn disable_rwc_int(&self)

Disable Region Wrap Condition detected interrupt enable

pub fn disable_rbe_int(&self)

Disable Region Bus Error interrupt enable

pub fn disable_rdm_int(&self)

Disable Region Digest Mismatch interrupt enable

pub fn disable_rhc_int(&self)

Disable Region Hash Completed interrupt enable

pub fn get_rsu_int_mask(&self) -> bool

Get Region Status Updated interrupt enable mask

pub fn get_rec_int_mask(&self) -> bool

Get Region End bit Condition Detected interrupt enable mask

pub fn get_rwc_int_mask(&self) -> bool

Get Region Wrap Condition detected interrupt enable mask

pub fn get_rbe_int_mask(&self) -> bool

Get Region Bus Error interrupt enable mask

pub fn get_rdm_int_mask(&self) -> bool

Get Region Digest Mismatch interrupt enable mask

pub fn get_rhc_int_mask(&self) -> bool

Get Region Hash Completed interrupt enable mask

pub fn get_rsu_int(&self) -> bool

Region Status Updated interrupt status

pub fn get_rec_int(&self) -> bool

Region End bit Condition Detected interrupt status

pub fn get_rwc_int(&self) -> bool

Region Wrap Condition detected interrupt status

pub fn get_rbe_int(&self) -> bool

Region Bus Error interrupt status

pub fn get_rdm_int(&self) -> bool

Region Digest Mismatch interrupt status

pub fn get_rhc_int(&self) -> bool

Region Hash Completed interrupt status

pub fn get_interrupt_status(&mut self) -> RegionInterrupt<I>

When reading the interrupt (ISR) register, it is cleared

This might result in unexpected results for example when sequentially trying to determine which interrupt triggered.

This is an alternative, return all the data from the register and parse later with the designated get_[name]_int functions.