Struct atsamd_hal::icm::RegionConfiguration

#[repr(C)]
pub struct RegionConfiguration(pub u32);

Region Configuration Structure

Follows C-structure conventions

Tuple Fields
0: u32
Implementations

impl RegionConfiguration

pub fn get_cdwbn(&self) -> bool

Compare Digest or Write Back Digest

true: in Compare Digest mode.

false: in Write Back Digest mode.

pub fn set_cdwbn(&mut self, value: bool)

Compare Digest or Write Back Digest

true: in Compare Digest mode.

false: in Write Back Digest mode.

pub fn get_wrap(&self) -> bool

Wrap Command

true: next region descriptor address loaded is DSCR.

false: the next region descriptor address is current + 0x10.

pub fn set_wrap(&mut self, value: bool)

Wrap Command

true: next region descriptor address loaded is DSCR.

false: the next region descriptor address is current + 0x10.

pub fn get_eom(&self) -> bool

End of Monitoring

true: the current descriptor terminates the Main List, WRAP bit has no effect.

false: the current descriptor does not terminate the monitoring.

pub fn set_eom(&mut self, value: bool)

End of Monitoring

true: the current descriptor terminates the Main List, WRAP bit has no effect.

false: the current descriptor does not terminate the monitoring.

pub fn get_rhien(&self) -> bool

Region Hash Completed Interrupt Disable

true: the RHC flag remains cleared even if the setting condition is met.

false: the RHC flag is set when the field NEXT = 0 is found in main or secondary list.

pub fn set_rhien(&mut self, value: bool)

Region Hash Completed Interrupt Disable

true: the RHC flag remains cleared even if the setting condition is met.

false: the RHC flag is set when the field NEXT = 0 is found in main or secondary list.

pub fn get_dmien(&self) -> bool

Digest Mismatch Interrupt Disable

true: the RBE flag remains cleared even if the setting condition is met.

false: the RBE flag is set when the hash value just calculated from the processed region differs from expected hash value.

pub fn set_dmien(&mut self, value: bool)

Digest Mismatch Interrupt Disable

true: the RBE flag remains cleared even if the setting condition is met.

false: the RBE flag is set when the hash value just calculated from the processed region differs from expected hash value.

pub fn get_beien(&self) -> bool

Bus Error Interrupt Disable

true: the flag remains cleared even if the setting condition is met.

false: the flag is set when an error is reported on the sysstem bus by the bus MATRIX.

pub fn set_beien(&mut self, value: bool)

Bus Error Interrupt Disable

true: the flag remains cleared even if the setting condition is met.

false: the flag is set when an error is reported on the sysstem bus by the bus MATRIX.

pub fn get_wcien(&self) -> bool

Wrap Condition Interrupt Disable

true: the RWC flag remains cleared even if the setting condition is met.

false: the RWC flag is set when the WRAP is encountered.

pub fn set_wcien(&mut self, value: bool)

Wrap Condition Interrupt Disable

true: the RWC flag remains cleared even if the setting condition is met.

false: the RWC flag is set when the WRAP is encountered.

pub fn get_ecien(&self) -> bool

End Bit Condition Interrupt Enable

true: the REC flag remains cleared even if the setting condition is met.

false: the REC flag is set when the descriptor having the EOM bit set is processed.

pub fn set_ecien(&mut self, value: bool)

End Bit Condition Interrupt Enable

true: the REC flag remains cleared even if the setting condition is met.

false: the REC flag is set when the descriptor having the EOM bit set is processed.

pub fn get_suien(&self) -> bool

Monitoring Status Updated Condition Interrupt Enable

true: the RSU flag remains cleared even if the condition is met.

false: the RSU flag is set when the corresponding descriptor is loaded from memory to ICM.

pub fn set_suien(&mut self, value: bool)

Monitoring Status Updated Condition Interrupt Enable

true: the RSU flag remains cleared even if the condition is met.

false: the RSU flag is set when the corresponding descriptor is loaded from memory to ICM.

pub fn get_procdly(&self) -> bool

Processing Delay

Allows setting short or long delay.

See ProcessingDelay

pub fn set_procdly(&mut self, value: bool)

Processing Delay

Allows setting short or long delay.

See ProcessingDelay

impl RegionConfiguration

pub fn set_algo(&mut self, value: icm_algorithm)

User SHA Algorithm

Allow setting this regions icm_algorithm.

pub fn get_algo(&mut self) -> icm_algorithm

User SHA Algorithm

Get the current user sha algorithm

pub fn reset_region_configuration_to_default(&mut self)

Reset the RegionConfiguration to default values

Useful if changing between hashing and monitoring, etc.