atsamd_hal::interrupt
Trait InterruptExt

pub trait InterruptExt: InterruptNumber + Copy {
    // Provided methods
    unsafe fn enable(self) { ... }
    fn disable(self) { ... }
    fn is_enabled(self) -> bool { ... }
    fn is_pending(self) -> bool { ... }
    fn pend(self) { ... }
    fn unpend(self) { ... }
    fn get_priority(self) -> Priority { ... }
    fn set_priority(self, prio: Priority) { ... }
    fn set_priority_with_cs(self, _cs: CriticalSection<'_>, prio: Priority) { ... }
}

An interrupt type that can be configured by the HAL to handle interrupts.

The PAC defined enum-level interrupts implement this trait.

Provided Methods

unsafe fn enable(self)

Enable the interrupt.

Safety

Do not enable any interrupt inside a critical section.

fn disable(self)

Disable the interrupt.

fn is_enabled(self) -> bool

Check if interrupt is enabled.

fn is_pending(self) -> bool

Check if interrupt is pending.

fn pend(self)

Set interrupt pending.

fn unpend(self)

Unset interrupt pending.

fn get_priority(self) -> Priority

Get the priority of the interrupt.

fn set_priority(self, prio: Priority)

Set the interrupt priority.

fn set_priority_with_cs(self, _cs: CriticalSection<'_>, prio: Priority)

Set the interrupt priority with an already-acquired critical section.

Equivalent to set_priority, except you pass a [CriticalSection] to prove you’ve already acquired a critical section. This prevents acquiring another one, which saves code size.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.