Macro atsamd_hal::define_pins

macro_rules! define_pins {
    ($(#[$topattr:meta])* struct $Type:ident,
     target_device: $target_device:ident,
     $( $(#[$attr:meta])* pin $name:ident = $pin_ident:ident),+ , ) => { ... };
}

This macro is a helper for defining a Pins type in a board support crate. This type is used to provide more meaningful aliases for the various GPIO pins for a given board.