Struct atsamd_hal::nvm::smart_eeprom::SmartEeprom

pub struct SmartEeprom<'a, T: SmartEepromState> { /* private fields */ }

Struct representing a SmartEEPROM instance.

It is generic over:

a lifetime of a stored Nvm reference

current state (Locked/Unlocked)

Implementations

impl<'a, T: SmartEepromState> SmartEeprom<'a, T>

pub unsafe fn get_slice<TP: SmartEepromPointableSize>(&self) -> &[TP]

Returns an immutable slice to SmartEEPROM mapped address space.

Underlying pointed type can be either u8, u16 or u32.

Safety

NVMCTRL.SEESTAT.BUSY register must be 0 before memory access can be performed.

pub fn get<TP: SmartEepromPointableSize>(&self, offset: usize, buffer: &mut [TP])

Retrieves data stored in SmartEEPROM at offset location and copies it to buffer.

Note: offset_in_bytes == sizeof::<TP>() * offset

pub fn iter<TP: SmartEepromPointableSize>(&'a self) -> SmartEepromIter<'a, TP>

Returns an iterator over SmartEEPROM address space.

impl<'a> SmartEeprom<'a, Unlocked>

pub unsafe fn get_mut_slice<TP: SmartEepromPointableSize>(
&mut self
) -> &mut [TP]

Returns a mutable slice to SmartEEPROM mapped address space.

Underlying pointed type can be either u8, u16 or u32.

Safety

NVMCTRL.SEESTAT.BUSY register must be 0 before memory access can be performed.

pub fn set<TP: SmartEepromPointableSize>(&mut self, offset: usize, buffer: &[TP])

Copies data in a buffer to SmartEEPROM at offset location

Note: offset_in_bytes == sizeof::<TP>() * offset

pub fn iter_mut<TP: SmartEepromPointableSize>(
&'a mut self
) -> SmartEepromIterMut<'a, TP>

Returns a mutable iterator over SmartEEPROM address space.

pub fn lock(self) -> SmartEeprom<'a, Locked>

Locks SmartEEPROM, allowing only to perform read operations

impl<'a> SmartEeprom<'a, Locked>

pub fn unlock(self) -> SmartEeprom<'a, Unlocked>

Unlocks SmartEEPROM, allowing to perform both read and write operations