Struct atsamd_hal::nvm::Nvm

pub struct Nvm { /* private fields */ }

Non-volatile memory controller

Implementations

impl Nvm

pub fn new(nvm: NVMCTRL) -> Self

Create a new NVM controller or handle failure from DSU

pub unsafe fn bank_swap(&mut self) -> !

Swap the flash banks. The processor will be reset, after which the inactive bank will become the active bank.

Safety

Ensure there is a working, memory safe program in place in the inactive bank before calling.

pub fn power_reduction_mode(&mut self, prm: PRM_A)

Set the power reduction mode

pub fn is_boot_protected(&self) -> bool

Check if the flash is boot protected

pub fn first_bank(&self) -> PhysicalBank

Get first bank

pub fn is_ready(&self) -> bool

Determine if the controller is busy writing or erasing

pub fn user_page(&self) -> Userpage

Read the user page

pub fn calibration_area(&self) -> CalibrationArea

Read the calibration area

pub fn temperatures_calibration_area(&self) -> TemperaturesCalibrationArea

Read the calibration area for temperatures

pub fn boot_protection(&mut self, protect: bool) -> Result<()>

Enable/disable boot protection on/off

Userpage’s NVM BOOT field defines a memory region that is supposed to be protected. NVMCTRL.STATUS.BOOTPROT is a readonly HW register populated on reset with a value from a userpage. By default, 0

pub unsafe fn write_from_slice(
&mut self,
destination_address: u32,
source_slice: &[u32]
) -> Result<()>

Write to flash memory from a slice

Safety

If destination_address is not word-aligned, an error is returned.

Using write()

pub unsafe fn write(
&mut self,
destination_address: u32,
source_address: u32,
words: u32
) -> Result<()>

Write to flash memory

Safety

If either destination_address or source_address are not word-aligned, an error is returned.

Writes to flash goes through the NVM controller NVM controller sets the PROGE/LOCKE flag if an error occurs, this is checked in manage_error_states to propagate the relevant error code

pub unsafe fn erase(
&mut self,
address: u32,
length: u32,
granularity: EraseGranularity
) -> Result<()>

Erase flash memory.

Unit of length depends on a chosen erasing granularity.

Safety

Erasing flash goes through the NVM controller NVM controller sets the PROGE/LOCKE flag if an error occurs, this is checked in manage_error_states to propagate the relevant error code.

pub fn smart_eeprom(&mut self) -> Result<'_>

Retrieve SmartEEPROM