Trait atsamd_hal::prelude::_embedded_hal_blocking_delay_DelayUs

pub trait _embedded_hal_blocking_delay_DelayUs<UXX> {
    fn delay_us(&mut self, us: UXX);
}

Microsecond delay

UXX denotes the range type of the delay time. UXX can be u8, u16, etc. A single type can implement this trait for different types of UXX.

Required Methods

fn delay_us(&mut self, us: UXX)

Pauses execution for us microseconds