Module atsamd_hal::pukcc

Public Key Cryptography Controller (PUKCC)

This module provides both low and high level abstractions for dealing with a PUKCC peripheral.

PUKCC consists of a set of functions (called services) hidden within a reserved region of memory. These functions usually make use of a separate piece of RAM to operate called CryptoRAM.

c_abi module contains raw structs and callable C-like function definitions. Pukcc wraps this low-level access API and exposes it in a safe manner.

WARNING!

This module has not been evaluated for correctness nor suitability for any use-case. Subtle implementation details may have catastrophic implications for the security of your cryptosystem, and users are advised to engage a cryptographer before making use of this module.

Modules

c_abi

Module that defines low-level constructs required for interaction with PUKCC

curves

Module that defines curves parametrizations

Structs

Pukcc

Struct representing a PUKCC peripheral.

SelfTestFailure

An error type representing failure modes a Pukcc::self_test service

Enums

CalculateCnsFailure

An error type representing failure modes for a Pukcc::zp_calculate_cns service

EcdsaSignFailure

An error type representing failure modes for a Pukcc::zp_ecdsa_sign_with_entropy and Pukcc::zp_ecdsa_sign_with_raw_k service

EcdsaSignatureVerificationFailure

An error type representing failure modes for a Pukcc::zp_ecdsa_verify_signature service

ExpModFailure

An error type representing failure modes for a Pukcc::modular_exponentiation service

ExpModMode

An enum describing available modes of operation of Pukcc::modular_exponentiation algoritm

ExpModWindowSize

An enum describing allowed, predefined window sizes for a calculation workspace in CryptoRAM for Pukcc::modular_exponentiation algorithm

ExpectedLengthError

An error type specifying an expected length of a slice in question

PukclReturnCode

An enum type that is a human readable representation of a low-level c_abi::PukclReturnCode type. Useful when used together with a Debug traits and formatters.

PukclReturnCodeInfo

PukclReturnCode nested enum subtype

PukclReturnCodeSevere

PukclReturnCode nested enum subtype

PukclReturnCodeWarning

PukclReturnCode nested enum subtype