Module atsamd_hal::pukcc::c_abi

Module that defines low-level constructs required for interaction with PUKCC

Structs

CRT

ClearFlags

Comp

CondCopy

CryptoRam

Slice wrapper that provides Rust-like access to CryptoRAM memory area

Div

ExpMod

FastCopy

Fill

Fmult

GCD

GF2NEcConvAffineToProjective

GF2NEcConvProjToAffine

GF2NEcDsaGenerateFast

GF2NEcDsaVerifyFast

GF2NEcPointIsOnCurve

GF2NEcRandomiseCoordinate

GF2NEccAddFast

GF2NEccDblFast

GF2NEccMulFast

PrimeGen

PukclHeader

PukclParams

PukclReturnCode

PukclStatus

RedMod

Rng

SelfTest

Smult

Square

Swap

ZpEcConvAffineToProjective

ZpEcConvProjToAffine

ZpEcDsaGenerateFast

ZpEcDsaQuickVerify

ZpEcDsaVerifyFast

ZpEcPointIsOnCurve

ZpEcRandomiseCoordinate

ZpEccAddFast

ZpEccAddSubFast

ZpEccDblFast

ZpEccMulFast

ZpEccQuickDualMulFast

Traits

CryptoRamSlice

Trait implemented for all &[u8] slices in order to provide a normalized way of downcasting pointers in a form accepted by PUKCC ABI

Service

Functions

wait_for_crypto_ram_clear_process⚠

Function that has to be called before using PUKCC peripheral

Type Definitions

nu1

pfu1

pu1

u1

u2

u4

Unions

PukclParamsUnion