atsamd_hal::sercom::i2c
Struct I2c

pub struct I2c<C: AnyConfig, D = NoneT> { /* private fields */ }

Abstraction over a I2C peripheral, allowing to perform I2C transactions.

Implementations

impl<C, S> I2c<C>
where C: AnyConfig<Sercom = S>, S: Sercom,

pub fn into_future<I>(self, _interrupts: I) -> I2cFuture<C>
where I: Binding<S::Interrupt, InterruptHandler<S>>,

Turn an I2c into an I2cFuture

impl<C: AnyConfig, D> I2c<C, D>

pub fn data_ptr(&self) -> *mut Word

Obtain a pointer to the DATA register. Necessary for DMA transfers.

pub fn read_flags(&self) -> Flags

Read the interrupt flags

pub fn clear_flags(&mut self, flags: Flags)

Clear interrupt status flags

pub fn enable_interrupts(&mut self, flags: Flags)

Enable interrupts for the specified flags.

pub fn disable_interrupts(&mut self, flags: Flags)

Disable interrupts for the specified flags.

pub fn read_status(&self) -> Status

Read the status flags

pub fn clear_status(&mut self, status: Status)

Clear the status flags

pub fn reconfigure<F>(&mut self, update: F)
where F: FnOnce(&mut SpecificConfig<C>),

Reconfigure the I2C peripheral.

Calling this method will temporarily disable the SERCOM peripheral, as some registers are enable-protected. This may interrupt any ongoing transactions.

use atsamd_hal::sercom::i2c::I2c;
i2c.reconfigure(|c| c.set_run_in_standby(false));

pub fn disable(self) -> C

Disable the I2C peripheral and return the underlying Config

impl<C: AnyConfig> I2c<C>

pub fn with_dma_channel<Chan: AnyChannel<Status = Ready>>( self, channel: Chan, ) -> I2c<C, Chan>

Attach a DMA channel to this I2c. Its embedded_hal::i2c::I2c implementation will use DMA to carry out its transactions.

impl<C, D, S> I2c<C, D>
where C: AnyConfig, D: AnyChannel<Status = S>, S: ReadyChannel,

pub fn take_dma_channel(self) -> (I2c<C, NoneT>, D)

Reclaim the DMA channel. Any subsequent I2C operations will no longer use DMA.

impl<C: AnyConfig> I2c<C>

pub fn init_dma_transfer(&mut self) -> Result<I2cBusReady, Error>

๐Deprecated since 0.19.0: Use I2c::with_dma_channel instead. You will have access to DMA-enabled embedded-hal implementations.

Initialize the bus to start receiving with DMA.

In reality, this function only checks whether or not the I2C bus is ready to accept a new transaction. A call to init_dma_transfer should immediately be followed by a call to send_with_dma or receive_with_dma.

// Assume `i2c` is a fully configured `I2c`, and `chan0` a fully configured `dmac::Channel`.
let token = i2c.init_dma_transfer()?;
i2c.send_with_dma(ADDRESS, token, buf_src, chan0);

pub fn receive_with_dma<Ch, B>( self, address: u8, _ready_token: I2cBusReady, buf: B, channel: Ch, ) -> Transfer<Channel<Ch::Id, Busy>, BufferPair<Self, B>>
where Ch: AnyChannel<Status = Ready>, B: Buffer<Beat = Word> + 'static,

๐Deprecated since 0.19.0: Use I2c::with_dma_channel instead. You will have access to DMA-enabled embedded-hal implementations.

Transform an I2c into a DMA Transfer) and start receiving into the provided buffer. The buffer length must be 255 bytes or shorter.

It is recommended that you check for errors after the transfer is complete by calling read_status.

pub fn send_with_dma<Ch, B>( self, address: u8, _ready_token: I2cBusReady, buf: B, channel: Ch, ) -> Transfer<Channel<Ch::Id, Busy>, BufferPair<B, Self>>
where Ch: AnyChannel<Status = Ready>, B: Buffer<Beat = Word> + 'static,

๐Deprecated since 0.19.0: Use I2c::with_dma_chahnnel instead. You will have access to DMA-enabled embedded-hal implementations.

Transform an I2c into a DMA Transfer) and start sending the provided buffer. The buffer length must be 255 bytes or shorter.

It is recommended that you check for errors after the transfer is complete by calling read_status.