Trait atsamd_hal::sercom::i2c::AnyConfig

pub trait AnyConfig: Is<Type = SpecificConfig<Self>> {
    type Sercom: Sercom;
    type Pads: PadSet<Sercom = Self::Sercom>;
}

Type class for all possible Config types

This trait uses the AnyKind trait pattern to create a type class for Config types. See the AnyKind documentation for more details on the pattern.

In addition to the normal, AnyKind associated types. This trait also copies the Sercom type, to make it easier to apply bounds to these types at the next level of abstraction.

Required Associated Types

type Sercom: Sercom

type Pads: PadSet<Sercom = Self::Sercom>