atsamd_hal::sercom::pad

Trait IoSet

pub trait IoSet: Sealed {
    type Order;
}

Type-level enum representing a SERCOM IOSET

The SAM D5x/E5x has particular sets of Pins that are allowed to be used together for each Sercom, and Pins from different sets cannot be used together. The valid combinations of Pins are called IOSETs (or IO SETs) in the datasheet. Additionally, some undocumented sets are used in commercially available boards (such as Adafruit’s PyGamer and Feather M4). This IoSet trait is used to constrain the various Pads types (spi::Pads, uart::Pads, and i2c::Pads) to only contain valid documented or undocumented sets of Pins.

See the type-level enum documentation for more details on the pattern. Typenum unsigned integers are used to make IoSets comparable

Required Associated Types

type Order