atsamd_hal::sercom::pad

Trait IsI2cPad

pub trait IsI2cPad: IsPad { }

Type class for Pins which can be used as I2C pads

This trait serves as a type class for Pins configured as I2C pads.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.