Trait atsamd_hal::sercom::pad::IsPad

pub trait IsPad: AnyPin {
    type Sercom: Sercom;
    type PadNum: PadNum;
}

Type class for Pins configured as SERCOM pads

This trait serves as both a type class for Pins configured as SERCOM pads and as a type-level function mapping each Pin type to its corresponding Sercom and PadNum.

Required Associated Types

type Sercom: Sercom

type PadNum: PadNum