atsamd_hal::sercom::spi

Module size

Define a trait to track the transaction Length, which represents the Config Size for SAMx5x chips

This module defines the Size trait for SAMx5x chips. These chips always operate in 32-bit extension mode and use the hardware LENGTH counter to track the length of each transaction, in bytes. See the Length documentation for more details.

Traits

GreaterThan4

Marker trait for transaction Lengths greater than four

Length

Type-level enum representing the SPI transaction length, in bytes

StaticLength

Marker trait for statically known transaction Lengths

Type Aliases

DynLength

Marker type for a run-time dynamic Length

Word

Type alias to recover the Word type from an implementation of Length