atsamd_hal::sercom::spi::size

Trait CharSize

pub trait CharSize: Sealed {
    type Word: 'static + Copy;

    const BITS: u8;
}

Type-level enum representing the SPI character size

This trait acts as both a type-level enum, forming a type class for character sizes, as well as a type-level function mapping the corresponding word size.

The SPI character size affects the word size for the embedded HAL traits. Eight-bit transactions use a u8 word, while nine-bit transactions use a u16 word.

Required Associated Constants

const BITS: u8

Register bit pattern for the corresponding CharSize

Required Associated Types

type Word: 'static + Copy

Word size for the character size

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.