Struct atsamd_hal::sercom::spi::Spi

pub struct Spi<C, A>where
    C: ValidConfig,
    A: Capability,{ /* private fields */ }

An enabled SPI peripheral that can perform transactions

See the impl_ehal documentation for details on the implementations of the embedded HAL traits, which vary based on Size and Capability.

Implementations

impl<C, A> Spi<C, A>where
C: ValidConfig,
A: Capability,

pub fn reconfigure(&mut self, update: impl FnOnce(&mut SpecificConfig<C>))

Update the SPI configuration.

Calling this method will temporarily disable the SERCOM peripheral, as some registers are enable-protected. This may interrupt any ongoing transactions.

pub fn enable_interrupts(&mut self, flags: Flags)

Enable interrupts for the specified flags

pub fn disable_interrupts(&mut self, flags: Flags)

Disable interrupts for the specified flags

pub fn read_flags(&self) -> Flags

Read the interrupt flags

pub fn clear_flags(&mut self, flags: Flags)

Clear the corresponding interrupt flags

Only the ERROR, SSL and TXC flags can be cleared.

Note: The implementation of of serial::Write::flush waits on and clears the TXC flag. Manually clearing this flag could cause it to hang indefinitely.

pub fn read_status(&self) -> Status

Read the error status flags

pub fn clear_status(&mut self, status: Status)

Clear the corresponding error status flags

pub fn read_flags_errors(&self) -> Result<Flags, Error>

Try to read the interrupt flags, but first check the error status flags.

pub unsafe fn read_data(&mut self) -> DataWidth

Read from the DATA register

Reading from the data register directly is unsafe, because it will clear the RXC flag, which could break assumptions made elsewhere in this module.

pub unsafe fn write_data(&mut self, data: DataWidth)

Write to the DATA register

Writing to the data register directly is unsafe, because it will clear the DRE flag, which could break assumptions made elsewhere in this module.

pub fn disable(self) -> C

Disable the SPI peripheral and return the Config struct