atsamd_hal::sercom::spi
Struct SpiFuture

pub struct SpiFuture<C, A, R = NoneT, T = NoneT>
where
    C: ValidConfig,
    A: Capability,
{ /* private fields */ }

async version of Spi.

Create this struct by calling Spi::into_future.

Implementations

impl<C, D, RxDma, TxDma> SpiFuture<C, D, RxDma, TxDma>
where C: ValidConfig, D: Capability, RxDma: AnyChannel<Status = ReadyFuture>, TxDma: AnyChannel<Status = ReadyFuture>,

pub fn take_dma_channels(self) -> (SpiFuture<C, D, NoneT, NoneT>, RxDma, TxDma)

Reclaim the DMA channels. Any subsequent SPI transaction will not use DMA.

impl<C, D, R, T> SpiFuture<C, D, R, T>
where C: ValidConfig, D: Receive, R: AnyChannel<Status = ReadyFuture>,

pub fn take_rx_channel(self) -> (SpiFuture<C, D, NoneT, T>, R)

Reclaim the Rx DMA channel. Any subsequent SPI transaction will not use DMA.

impl<C, D, R, T> SpiFuture<C, D, R, T>
where C: ValidConfig, D: Capability, T: AnyChannel<Status = ReadyFuture>,

pub fn take_tx_channel(self) -> (SpiFuture<C, D, R, NoneT>, T)

Reclaim the DMA channel. Any subsequent SPI transaction will not use DMA.

impl<C, S, R> SpiFuture<C, Rx, R, NoneT>
where C: ValidConfig<Sercom = S>, C::Word: PrimInt + AsPrimitive<DataWidth>, DataWidth: AsPrimitive<C::Word>, S: Sercom,

pub fn with_rx_dma_channel<Chan: AnyChannel<Status = ReadyFuture>>( self, rx_channel: Chan, ) -> SpiFuture<C, Rx, Chan, NoneT>

Add a DMA channel for receiving transactions

impl<C, S, T> SpiFuture<C, Tx, NoneT, T>
where C: ValidConfig<Sercom = S>, C::Word: PrimInt + AsPrimitive<DataWidth>, DataWidth: AsPrimitive<C::Word>, S: Sercom,

pub fn with_tx_dma_channel<Chan: AnyChannel<Status = ReadyFuture>>( self, tx_channel: Chan, ) -> SpiFuture<C, Tx, NoneT, Chan>

Add a DMA channel for receiving transactions

impl<C, S, R, T> SpiFuture<C, Duplex, R, T>
where C: ValidConfig<Sercom = S>, C::Word: PrimInt + AsPrimitive<DataWidth>, DataWidth: AsPrimitive<C::Word>, S: Sercom,

pub fn with_dma_channels<ChanRx, ChanTx>( self, rx_channel: ChanRx, tx_channel: ChanTx, ) -> SpiFuture<C, Duplex, ChanRx, ChanTx>
where ChanRx: AnyChannel<Status = ReadyFuture>, ChanTx: AnyChannel<Status = ReadyFuture>,

Add a DMA channel for receiving transactions

impl<P, M, Z, D, R, T, S> SpiFuture<Config<P, M, Z>, D, R, T>
where P: ValidPads, M: OpMode, Z: Size + 'static, Config<P, M, Z>: ValidConfig<Sercom = S>, D: Transmit, S: Sercom, Z::Word: PrimInt + AsPrimitive<DataWidth> + Beat, DataWidth: AsPrimitive<Z::Word>, T: AnyChannel<Status = ReadyFuture>,

pub async fn write_dma(&mut self, words: &[Z::Word]) -> Result<usize, Error>

Write words from a buffer asynchronously, using DMA.

impl<P, M, S, C, D, R, T> SpiFuture<Config<P, M, C>, D, R, T>
where Config<P, M, C>: ValidConfig<Sercom = S>, S: Sercom, P: ValidPads, M: MasterMode, C: Size + 'static, C::Word: PrimInt + AsPrimitive<DataWidth> + Beat, D: Capability, DataWidth: AsPrimitive<C::Word>, R: AnyChannel<Status = ReadyFuture>, T: AnyChannel<Status = ReadyFuture>,

pub async fn read_dma_master( &mut self, words: &mut [C::Word], ) -> Result<(), Error>

Read words into a buffer asynchronously, using DMA.

impl<C, D> SpiFuture<C, D, NoneT, NoneT>
where C: ValidConfig, D: Capability,

pub fn free(self) -> Spi<C, D>

Return the underlying Spi.

impl<P, M, C, D, R, S> SpiFuture<Config<P, M, C>, D, R, NoneT>
where Config<P, M, C>: ValidConfig<Sercom = S>, P: ValidPads, M: OpMode, C: Size + 'static, C::Word: PrimInt + AsPrimitive<DataWidth>, D: Transmit, DataWidth: AsPrimitive<C::Word>, S: Sercom,

pub async fn write(&mut self, words: &[C::Word]) -> Result<(), Error>

Write words from a buffer asynchronously, word by word

impl<P, M, C, D, T, S> SpiFuture<Config<P, M, C>, D, NoneT, T>
where Config<P, M, C>: ValidConfig<Sercom = S>, P: ValidPads, M: MasterMode, C: Size + 'static, C::Word: PrimInt + AsPrimitive<DataWidth>, D: Receive, DataWidth: AsPrimitive<C::Word>, S: Sercom,

pub async fn read(&mut self, buffer: &mut [C::Word]) -> Result<(), Error>

Read words into a buffer asynchronously, word by word.

Since we are using a Duplex SpiFuture, we need to send a word simultaneously while receiving one. This no-op word is configurable via the nop_word method.

impl<P, M, C, S, D, R, T> SpiFuture<Config<P, M, C>, D, R, T>
where Config<P, M, C>: ValidConfig<Sercom = S>, P: ValidPads, M: OpMode, C: Size + 'static, C::Word: PrimInt + AsPrimitive<DataWidth>, DataWidth: AsPrimitive<C::Word>, D: Capability, S: Sercom,

pub async fn transfer_word_in_place( &mut self, to_send: C::Word, ) -> Result<C::Word, Error>

Read and write a single word to the bus simultaneously.