atsamd_hal::sercom::spi

Trait AtomicSize

pub trait AtomicSize: Size { }

Marker trait for transaction Sizes that can be completed in a single read or write of the DATA register

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.