Trait atsamd_hal::sercom::spi::ValidConfig

pub trait ValidConfig: AnyConfig { }

Marker trait for valid SPI Configurations

A functional SPI peripheral must have, at a minimum, an SCLK pad and either a Data In or a Data Out pad. Dependeing on the OpMode, an SS pad may also be required.

The ValidConfig trait is implemented only for valid combinations of Pads and OpMode. No Config is valid if the SCK pad is NoneT or if both the Data In and Data Out pads are NoneT. When in Master OpMode, the SS pad must be NoneT, while in MasterHWSS or Slave OpMode, the SS pad must be SomePad.