Trait atsamd_hal::sercom::spi_future::ControlSS

pub trait ControlSS: OptionalPin {
    fn assert(&mut self);
    fn deassert(&mut self);
}

Trait used to control the SS line during an SpiFuture transaction

Required Methods

fn assert(&mut self)

If an SS pin is present, assert it by bringing it low

fn deassert(&mut self)

If an SS pin is present, deassert it by bringing it high