atsamd_hal::sercom::uart
Struct UartFuture

pub struct UartFuture<C, D, R = NoneT, T = NoneT>
where
    C: ValidConfig,
    D: Capability,
{ /* private fields */ }

async version of a Uart.

Create this struct by calling Uart::into_future.

Implementations

impl<C, D, R, T> UartFuture<C, D, R, T>
where C: ValidConfig, D: Capability, R: AnyChannel<Status = ReadyFuture>,

pub fn take_rx_channel(self) -> (UartFuture<C, D, NoneT, T>, R)

Reclaim the RX DMA channel. Subsequent RX operations will no longer use DMA.

impl<C, D, R, T> UartFuture<C, D, R, T>
where C: ValidConfig, D: Capability, T: AnyChannel<Status = ReadyFuture>,

pub fn take_tx_channel(self) -> (UartFuture<C, D, R, NoneT>, T)

Reclaim the TX DMA channel. Subsequent TX operations will no longer use DMA.

impl<C, D, S, R, T> UartFuture<C, D, R, T>
where C: ValidConfig<Sercom = S>, C::Word: Beat, D: Receive, S: Sercom + 'static, DataReg: AsPrimitive<C::Word>, R: AnyChannel<Status = ReadyFuture>,

pub async fn read(&mut self, words: &mut [C::Word]) -> Result<(), Error>

Read the specified number of Words into a buffer using DMA.

impl<C, D, S, R, T> UartFuture<C, D, R, T>
where C: ValidConfig<Sercom = S>, C::Word: Beat, D: Transmit, S: Sercom + 'static, T: AnyChannel<Status = ReadyFuture>,

pub async fn write(&mut self, words: &[C::Word]) -> Result<(), Error>

Write words from a buffer asynchronously, using DMA.

impl<C, R, T> UartFuture<C, Duplex, R, T>
where C: ValidConfig,

pub fn split( self, ) -> (UartFuture<C, RxDuplex, R, NoneT>, UartFuture<C, TxDuplex, NoneT, T>)

Split the UartFuture into RxDuplexand TxDuplex halves.

pub fn join( rx: UartFuture<C, RxDuplex, R, NoneT>, tx: UartFuture<C, TxDuplex, NoneT, T>, ) -> Self

Join RxDuplex and TxDuplex halves back into a full UartFuture<C, Duplex>

impl<C, D> UartFuture<C, D, NoneT, NoneT>
where C: ValidConfig, D: SingleOwner,

pub fn free(self) -> Uart<C, D>

Return the underlying Uart.

impl<C, D, S, R, T> UartFuture<C, D, R, T>
where C: ValidConfig<Sercom = S>, D: Receive, S: Sercom, DataReg: AsPrimitive<C::Word>,

pub fn with_rx_dma_channel<Chan: AnyChannel<Status = ReadyFuture>>( self, rx_channel: Chan, ) -> UartFuture<C, D, Chan, T>

Use a DMA channel for receiving words on the RX line

pub async fn read_word(&mut self) -> Result<C::Word, Error>

Read a single Word from the UART.

impl<C, D, S, T> UartFuture<C, D, NoneT, T>
where C: ValidConfig<Sercom = S>, D: Receive, S: Sercom, DataReg: AsPrimitive<C::Word>,

pub async fn read( &mut self, buffer: &mut [C::Word], ) -> Result<(), (Error, usize)>

Read the specified number of Words into a buffer, word by word.

In case of an error, returns Err(Error, usize) where the usize represents the number of valid words read before the error occured.

impl<C, D, S, R, T> UartFuture<C, D, R, T>
where C: ValidConfig<Sercom = S>, D: Transmit, S: Sercom,

pub fn with_tx_dma_channel<Chan: AnyChannel<Status = ReadyFuture>>( self, tx_channel: Chan, ) -> UartFuture<C, D, R, Chan>

Use a DMA channel for sending words on the TX line

pub async fn write_word(&mut self, word: C::Word)

Write a single Word to the UART.

impl<C, D, S, R> UartFuture<C, D, R, NoneT>
where C: ValidConfig<Sercom = S>, D: Transmit, S: Sercom,

pub async fn write(&mut self, buffer: &[C::Word])

Write the specified number of Words from a buffer to the UART, word by word.