atsamd_hal::sercom::uart

Trait AnyConfig

pub trait AnyConfig: Sealed + Is<Type = SpecificConfig<Self>> {
    type Sercom: Sercom;
    type Pads: ValidPads<Sercom = Self::Sercom>;
    type Word: 'static + PrimInt + AsPrimitive<DataReg>;
    type CharSize: CharSize<Word = Self::Word>;
}

Type class for all possible Config types

This trait uses the AnyKind trait pattern to create a [type class] for Config types. See the AnyKind documentation for more details on the pattern.

In addition to the normal, AnyKind associated types. This trait also copies the Sercom and Word types, to make it easier to apply bounds to these types at the next level of abstraction.

Required Associated Types

type Sercom: Sercom

type Pads: ValidPads<Sercom = Self::Sercom>

type Word: 'static + PrimInt + AsPrimitive<DataReg>

type CharSize: CharSize<Word = Self::Word>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.