Trait atsamd_hal::sercom::uart::RxpoTxpo

pub trait RxpoTxpo {
    const RXPO: u8;
    const TXPO: u8;
}

Configure the RXPO and TXPO fields based on a set of Pads

According to the datasheet, the RXPO and TXPO values specify which SERCOM pads are used for various functions. Moreover, depending on which pads are actually in use, only certain combinations of these values make sense and are valid.

This trait is implemented for valid, four-tuple combinations of OptionalPadNums. Those implementations are then lifted to the corresponding Pads types.

To satisfy this trait, the combination of OptionalPadNums must specify PadNum for at least one of RX and TX. Furthermore, no two PadNums can conflict.

Required Associated Constants

const RXPO: u8

RXPO field value

const TXPO: u8

RXPO field value