Trait atsamd_hal::sercom::uart::ValidPads

pub trait ValidPads: PadSet + RxpoTxpo {
    type Capability: Capability;
}

Marker trait for valid sets of Pads

This trait labels sets of Pads that satisfy the RxpoTxpo trait. It guarantees to the Config struct that this set of Pads can be configured through those traits.

Required Associated Types

type Capability: Capability