Type Definition atsamd_hal::sercom::uart::PadsFromIds

pub type PadsFromIds<S, RX = NoneT, TX = NoneT, RTS = NoneT, CTS = NoneT> = Pads<S, <RX as GetOptionalPad<S>>::Pad, <TX as GetOptionalPad<S>>::Pad, <RTS as GetOptionalPad<S>>::Pad, <CTS as GetOptionalPad<S>>::Pad>;

Define a set of Pads using PinIds instead of Pins

In some cases, it is more convenient to specify a set of Pads using PinIds rather than Pins. This alias makes it easier to do so.

The first type parameter is the Sercom, while the remaining four are effectively OptionalPinIds representing the corresponding type parameters of Pads, i.e. RX, TX, RTS & CTS. Each of the remaining type parameters defaults to NoneT.

use atsamd_hal::pac::Peripherals;
use atsamd_hal::gpio::{PA08, PA09, Pins};
use atsamd_hal::sercom::{Sercom0, uart};
use atsamd_hal::typelevel::NoneT;

pub type Pads = uart::PadsFromIds<Sercom0, PA08, PA09>;

pub fn create_pads() -> Pads {
    let peripherals = Peripherals::take().unwrap();
    let pins = Pins::new(peripherals.PORT);
    uart::Pads::default().rx(pins.pa09).tx(pins.pa08)
}