Module atsamd_hal::sercom::v1

Version 1 of the SERCOM module

This module retains the previous SERCOM API. The pads module provides a compatibility shim that uses the new v2::pad module to implement the old API. This API will eventually be deprecated and removed.

Re-exports

pub use pads::*;

Modules

i2c

pads
Version 1 of the SERCOM pads module

spi

uart

Structs

I2CMaster0
Represents the Sercom instance configured to act as an I2C Master. The embedded_hal blocking I2C traits are implemented by this instance.

I2CMaster1
Represents the Sercom instance configured to act as an I2C Master. The embedded_hal blocking I2C traits are implemented by this instance.

I2CMaster2
Represents the Sercom instance configured to act as an I2C Master. The embedded_hal blocking I2C traits are implemented by this instance.

I2CMaster3
Represents the Sercom instance configured to act as an I2C Master. The embedded_hal blocking I2C traits are implemented by this instance.

I2CMaster4
Represents the Sercom instance configured to act as an I2C Master. The embedded_hal blocking I2C traits are implemented by this instance.

I2CMaster5
Represents the Sercom instance configured to act as an I2C Master. The embedded_hal blocking I2C traits are implemented by this instance.

I2CMaster6
Represents the Sercom instance configured to act as an I2C Master. The embedded_hal blocking I2C traits are implemented by this instance.

I2CMaster7
Represents the Sercom instance configured to act as an I2C Master. The embedded_hal blocking I2C traits are implemented by this instance.

Padout
A pad mapping configuration for the SERCOM in SPI master mode.

Padout
A pad mapping configuration for the SERCOM in UART mode.

RxPadout
A pad mapping configuration for the receiving half of the SERCOM in UART mode.

SPIMaster0
SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

SPIMaster1
SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

SPIMaster2
SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

SPIMaster3
SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

SPIMaster4
SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

SPIMaster5
SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

SPIMaster6
SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

SPIMaster7
SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

TxPadout
A pad mapping configuration for the transmitting half of the SERCOM in UART mode.

UART0
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART0Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART0Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART1
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART1Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART1Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART2
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART2Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART2Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART3
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART3Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART3Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART4
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART4Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART4Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART5
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART5Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART5Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART6
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART6Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART6Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART7
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART7Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART7Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

Enums

Error

I2CError

Traits

DipoDopo
The DipoDopo trait defines a way to get the data in and data out pin out values for a given SPIMasterXPadout configuration. You should not implement this trait for yourself; only the implementations in the sercom module make sense.

RxpoTxpo
The RxpoTxpo trait defines a way to get the data in and data out pin out values for a given UARTXPadout configuration. You should not implement this trait for yourself; only the implementations in the sercom module make sense.

Type Definitions

SPIMaster0Padout

SPIMaster1Padout

SPIMaster2Padout

SPIMaster3Padout

SPIMaster4Padout

SPIMaster5Padout

SPIMaster6Padout

SPIMaster7Padout

UART0Padout

UART0RxPadout

UART0TxPadout

UART1Padout

UART1RxPadout

UART1TxPadout

UART2Padout

UART2RxPadout

UART2TxPadout

UART3Padout

UART3RxPadout

UART3TxPadout

UART4Padout

UART4RxPadout

UART4TxPadout

UART5Padout

UART5RxPadout

UART5TxPadout

UART6Padout

UART6RxPadout

UART6TxPadout

UART7Padout

UART7RxPadout

UART7TxPadout