Trait atsamd_hal::sercom::v1::pads::PadPin

pub trait PadPin<P>: Sealed {
    fn into_pad(self, port: &mut Port) -> P;
}

The PadPin trait makes it more ergonomic to convert a pin into a Sercom pad. You should not implement this trait for yourself; only the implementations in the sercom module make sense.

Required methods

fn into_pad(self, port: &mut Port) -> P

Loading content...