Struct atsamd_hal::sercom::v1::I2CMaster0

pub struct I2CMaster0<P0, P1> where
    P0: CompatiblePad<Sercom = SERCOM0, PadNum = Pad0>,
    P1: CompatiblePad<Sercom = SERCOM0, PadNum = Pad1>,  { /* fields omitted */ }

Represents the Sercom instance configured to act as an I2C Master. The embedded_hal blocking I2C traits are implemented by this instance.

Implementations

impl<P0, P1> I2CMaster0<P0, P1> where
P0: CompatiblePad<Sercom = SERCOM0, PadNum = Pad0>,
P1: CompatiblePad<Sercom = SERCOM0, PadNum = Pad1>,

pub fn new<F: Into<Hertz>>(
clock: &Sercom0CoreClock,
freq: F,
sercom: SERCOM0,
pm: &mut PM,
sda: P0,
scl: P1
) -> Self

Configures the sercom instance to work as an I2C Master. The clock is obtained via the GenericClockGenerator type. freq specifies the bus frequency to use for I2C communication. There are typically a handful of values that tend to be supported; standard mode is 100.khz(), full speed mode is 400.khz(). The hardware in the atsamd device supports fast mode at 1.mhz() and fast mode, but there may be additional hardware configuration missing from the current software implementation that prevents that from working as-written today.

let mut i2c = I2CMaster3::new(
    &clocks.sercom3_core(&gclk0).unwrap(),
    400.khz(),
    p.device.SERCOM3,
    &mut p.device.PM,
    // Metro M0 express has I2C on pins PA22, PA23
    pins.pa22.into_pad(&mut pins.port),
    pins.pa23.into_pad(&mut pins.port),
);

pub fn free(self) -> (P0, P1, SERCOM0)

Breaks the sercom device up into its constituent pins and the SERCOM instance. Does not make any changes to power management.