Trait atsamd_hal::sercom::v1::DipoDopo

pub trait DipoDopo {
    const DIPO: u8;
    const DOPO: u8;
    fn dipo_dopo(&self) -> (u8, u8) { ... }
}

The DipoDopo trait defines a way to get the data in and data out pin out values for a given SPIMasterXPadout configuration. You should not implement this trait for yourself; only the implementations in the sercom module make sense.

Associated Constants

const DIPO: u8

const DOPO: u8

Loading content...
Provided methods

fn dipo_dopo(&self) -> (u8, u8)

Loading content...