Trait atsamd_hal::sercom::v1::RxpoTxpo

pub trait RxpoTxpo {
    const RXPO: u8;
    const TXPO: u8;
    fn rxpo_txpo(&self) -> (u8, u8) { ... }
}

The RxpoTxpo trait defines a way to get the data in and data out pin out values for a given UARTXPadout configuration. You should not implement this trait for yourself; only the implementations in the sercom module make sense.

Associated Constants

const RXPO: u8

const TXPO: u8

Loading content...
Provided methods

fn rxpo_txpo(&self) -> (u8, u8)

Loading content...