Module atsamd_hal::sercom::v1::uart

Structs

Padout
A pad mapping configuration for the SERCOM in UART mode.

RxPadout
A pad mapping configuration for the receiving half of the SERCOM in UART mode.

TxPadout
A pad mapping configuration for the transmitting half of the SERCOM in UART mode.

UART0
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART0Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART0Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART1
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART1Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART1Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART2
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART2Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART2Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART3
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART3Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART3Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART4
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART4Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART4Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

UART5
UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

UART5Rx
The receiving half of the corresponding UARTX instance (as returned by UARTX::split)

UART5Tx
The transmitting half of the corresponding UARTX instance (as returned by UARTX::split)

Traits

RxpoTxpo
The RxpoTxpo trait defines a way to get the data in and data out pin out values for a given UARTXPadout configuration. You should not implement this trait for yourself; only the implementations in the sercom module make sense.

Type Definitions

UART0Padout

UART0RxPadout

UART0TxPadout

UART1Padout

UART1RxPadout

UART1TxPadout

UART2Padout

UART2RxPadout

UART2TxPadout

UART3Padout

UART3RxPadout

UART3TxPadout

UART4Padout

UART4RxPadout

UART4TxPadout

UART5Padout

UART5RxPadout

UART5TxPadout