Struct atsamd_hal::sercom::v1::uart::Padout

pub struct Padout<S, RX, TX, RTS, CTS> where
    S: Sercom,  { /* fields omitted */ }

A pad mapping configuration for the SERCOM in UART mode.

This type can only be constructed using the From implementations in this module, which are restricted to valid configurations.

Defines which sercom pad is mapped to which UART function.

Implementations

impl<S, RX, TX, RTS, CTS> Padout<S, RX, TX, RTS, CTS> where
S: Sercom,

pub fn split(self) -> (TxPadout<S, TX, RTS>, RxPadout<S, RX, CTS>)

Splits the padout into transmit and receive halves

pub fn join(tx: TxPadout<S, TX, RTS>, rx: RxPadout<S, RX, CTS>) -> Self

Combines transmit and receive halves back into a duplex padout