Struct atsamd_hal::sercom::v1::uart::UART5

pub struct UART5<RX, TX, RTS, CTS> { /* fields omitted */ }

UARTX represents the corresponding SERCOMX instance configured to act in the role of a UART Master. Objects of this type implement the HAL serial::Read, serial::Write traits.

This type is generic over any valid pad mapping where there is a defined “receive pin out transmit pin out” implementation.

Implementations

impl<RX, TX, RTS, CTS> UART5<RX, TX, RTS, CTS>

pub fn new<F: Into<Hertz>, T: Into<Padout<Sercom5, RX, TX, RTS, CTS>>>(
clock: &Sercom5CoreClock,
freq: F,
sercom: SERCOM5,
pm: &mut PM,
padout: T
) -> UART5<RX, TX, RTS, CTS> where
Padout<Sercom5, RX, TX, RTS, CTS>: RxpoTxpo,

Power on and configure SERCOMX to work as a UART Master operating with the specified frequency. The padout specifies which pins are bound to the RX, TX and optionally RTS and CTS functions.

You can use any tuple of two or four SercomXPadY instances for which there exists a From implementation for UARTXPadout.

pub fn free(self) -> (Padout<Sercom5, RX, TX, RTS, CTS>, SERCOM5)

pub fn split(self) -> (UART5Tx<TX, RTS>, UART5Rx<RX, CTS>)

Splits the UART into transmit and receive halves

pub fn join(tx: UART5Tx<TX, RTS>, rx: UART5Rx<RX, CTS>) -> Self

Combines transmit and receive halves back into a duplex UART

pub fn intenset<F>(&mut self, f: F) where
F: FnOnce(&mut W),

pub fn intenclr<F>(&mut self, f: F) where
F: FnOnce(&mut W),

pub fn flags(&self) -> R