Module atsamd_hal::sercom::v2

Version 2 of the SERCOM module

This module provides a new API for the SERCOM peripherals. So far, only the pad and spi modules have been updated, but it is expected that the uart, and i2c modules will eventually receive updates as well.

Re-exports

pub use pad::*;

Modules

impl_pad
Implementations of the IsPad, GetPad and InIoSet traits

pad
Define a SERCOM pad type

spi
Use the SERCOM peripheral for SPI transactions

spi_future
A Future-like interface for SPI transactions

Traits

Sercom
Type-level enum representing a Serial Communication Interface (SERCOM)

Type Definitions

Sercom0
Type alias for the corresponding SERCOM instance

Sercom1
Type alias for the corresponding SERCOM instance

Sercom2
Type alias for the corresponding SERCOM instance

Sercom3
Type alias for the corresponding SERCOM instance

Sercom4
Type alias for the corresponding SERCOM instance

Sercom5
Type alias for the corresponding SERCOM instance