Struct atsamd_hal::sercom::v2::spi::Config

pub struct Config<P, M = Master, L = U1> where
    P: ValidPads,
    M: OpMode,
    L: Length,  { /* fields omitted */ }

A configurable, disabled SPI peripheral

This struct represents a configurable SPI peripheral in its disabled state. It is generic over the set of Pads, OpMode and transaction Length. Upon creation, the Config takes ownership of the Sercom and resets it, returning it configured as an SPI peripheral in Master OpMode with Length U1.

Config uses a builder-pattern API to configure the peripheral, culminating in a call to enable, which consumes the Config and returns an enabled Spi peripheral. The enable function is restricted to ValidConfigs.

Implementations

impl<P: ValidPads> Config<P>

pub fn new(
mclk: &MCLK,
sercom: P::Sercom,
pad_map: P,
freq: impl Into<Hertz>
) -> Self

Create a new Config in the default configuration

This function will enable the corresponding APB clock, reset the Sercom peripheral, and return a Config in the default configuration, Master OpMode with Length U1. Config takes ownership of the Sercom and Pads.

Users must configure GCLK manually. The freq parameter represents the GCLK frequency for this Sercom instance.

impl<P, M, L> Config<P, M, L> where
P: ValidPads,
M: OpMode,
L: Length,

pub fn reset(self) -> Config<P>

Trigger the Sercom’s SWRST and return a Config in the default configuration.

pub unsafe fn sercom(&self) -> &P::Sercom

Obtain a reference to the PAC SERCOM struct

Directly accessing the SERCOM could break the invariants of the type-level tracking in this module, so it is unsafe.

pub fn free(self) -> (P::Sercom, P)

Consume the Config, reset the peripheral, and return the Sercom and Pads

pub fn op_mode<M2: OpMode>(self) -> Config<P, M2, L>

Change the OpMode

pub fn length<L2: Length>(self) -> Config<P, M, L2>

Change the transaction Length

To use a run-time dynamic length, set the Length type to DynLength and then use the get_dyn_length and set_dyn_length methods.

pub fn cpol(self, cpol: Polarity) -> Self

Change the clock polarity

pub fn cpha(self, cpha: Phase) -> Self

Change the clock phase

pub fn spi_mode(self, mode: Mode) -> Self

Change the SPI mode (clock polarity & phase)

pub fn msb_first(self, msb_first: bool) -> Self

Change the bit order of transmission (MSB/LSB first)

This only affects the order of bits within each byte. Bytes are always transferred in little endian order from the 32-bit DATA register.

pub fn baud<B: Into<Hertz>>(self, baud: B) -> Self

Set the baud rate

This function will calculate the best BAUD register setting based on the stored GCLK frequency and desired baud rate. The maximum baud rate is half the GCLK frequency. The minimum baud rate is the GCLK frequency / 512. Values outside this range will saturate at the extremes.

pub fn immediate_overflow_notification(&mut self, set: bool)

Control the buffer overflow notification

If set to true, an Error::Overflow will be issued as soon as an overflow occurs. Otherwise, it will not be issued until its place within the data stream.

pub fn run_in_standby(&mut self, set: bool)

Run in standby mode

When set, the SPI peripheral will run in standby mode. See the datasheet for more details.

pub fn enable_interrupts(&mut self, flags: Flags)

Enable interrupts for the specified flags

pub fn disable_interrupts(&mut self, flags: Flags)

Disable interrupts for the specified flags

pub fn enable(self) -> Spi<Self> where
Self: ValidConfig,

Enable the SPI peripheral

SPI transactions are not possible until the peripheral is enabled. This function is limited to ValidConfigs.

impl<P, M> Config<P, M, DynLength> where
P: ValidPads,
M: OpMode,

pub fn get_dyn_length(&self) -> u8

Return the current transaction length

Read the LENGTH register to determine the current transaction length

pub fn set_dyn_length(&mut self, length: u8)

Set the transaction length

Write the LENGTH register to set the transaction length. Panics if the length is zero.