Struct atsamd_hal::sercom::v2::spi::Spi

pub struct Spi<C: ValidConfig> { /* fields omitted */ }

An enabled SPI peripheral that can perform transactions

As noted in the OpMode and Length traits, this module chooses to always operate in 32-bit extension mode and uses the LENGTH counter to set the number of bytes in each transaction, from 1 to 255. In 32-bit extension mode, transaction Lengths of four bytes or fewer can be completed in a single read or write of the DATA register. Longer transactions require multiple reads or writes of the DATA register.

Embedded HAL traits

The embedded HAL FullDuplex trait is non-blocking, which has consequences for different transaction Lengths. Transaction Lengths of 1-4 bytes fit the FullDuplex model well, because they can return WouldBlock until it is possible to perform a single read or write that completes the transaction. The FullDuplex word type is u8 for 1-byte transactions, u16 for 2-byte transactions and u32 for 3- and 4-byte transactions.

For half-duplex, non-blocking transactions, this type implements the serial Read and Write traits. They are only implemented for transaction Lengths <= 4 and only when the Pads are not Tx+Rx, so they don’t compete with FullDuplex.

Longer Lengths require multiple reads or writes, which does not fit the FullDuplex model very well. When returning WouldBlock, the implementation would have to store some sort of internal state to track the progress of an on-going transaction. As a consequence, FullDuplex is only implemented for Lengths of 1-4.

However, the blocking::spi traits are different, because they can always execute a transaction to completion before returning. For Lengths of 1-4 bytes, these traits are implemented using an approach very similar to the Default implementations. Longer Lengths use a custom implementation of Transfer and Write and accept slices of u8 with a length matching the transaction Length.

When using a DynLength, Spi only implements Transfer and Write, using the same approach as long transaction Lengths.

For a non-blocking alternative that can be used to transfer slices, see the SpiFuture type.

Further documentation

The individual implementations of the embedded HAL traits are well documented. Unfortunately, however, rustdoc does not produce the most visually scannable documentation, especially when using typenum. Here are some tips for finding more information in the impl blocks below:

FullDuplex uses a single blanket implementation for Lengths of 1-4

Read has two different implementations, one for Slave OpMode and another for MasterModes

Write uses a single blanket implementation, like FullDuplex. However, watch out for ambiguity between serial::Write, blocking::serial::Write and blocking::spi::Write.

blocking::serial::Write uses the blocking::serial::write::Default implementation

blocking::spi::Transfer uses custom implementations. The implementations for Lengths of 1-4 are generated by macro, so there are four different implementations for U1 - U4. This is responsible for most of the visual noise in the docs. Lengths greater than four use a blanket implementation. DynLength uses its own implementation with an added run-time check of the transaction length.

blocking::spi::Write takes the same approach as Transfer.

blocking::spi::WriteIter is implemented via macro for Lengths 1-4, but it is not implemented for longer Lengths or DynLength.

Implementations

impl<C: ValidConfig> Spi<C>

pub unsafe fn sercom(&self) -> &C::Sercom

Obtain a reference to the PAC SERCOM struct

Directly accessing the SERCOM could break the invariants of the type-level tracking in this module, so it is unsafe.

pub fn reconfigure<F>(&mut self, update: F) where
F: FnOnce(SpecificConfig<C>) -> SpecificConfig<C>,

Update the SPI configuration.

Calling this method will temporarily disable the SERCOM peripheral, as some registers are enable-protected. This may interrupt any ongoing transactions.

pub unsafe fn length<L: Length>(self) -> Spi<Config<C::Pads, C::OpMode, L>> where
Config<C::Pads, C::OpMode, L>: ValidConfig,

Change the transaction Length

Changing the transaction Length while is enabled is permissible but unsafe. If you have sent or received any bytes at the current Length, you must wait for a TXC flag before changing to a new Length.

pub fn enable_interrupts(&mut self, flags: Flags)

Enable interrupts for the specified flags

pub fn disable_interrupts(&mut self, flags: Flags)

Disable interrupts for the specified flags

pub fn read_flags(&self) -> Flags

Read the interrupt status flags

pub fn clear_flags(&mut self, flags: Flags)

Clear interrupt status flags

Setting the ERROR, SSL or TXC flag will clear the interrupt. Clearing any flag will have no effect. This function has no effect on the DRE or RXC flags.

Warning: The implementation of of Write::flush waits on and clears the TXC flag. Manually clearing this flag could cause it to hang indefinitely.

pub fn read_errors(&self) -> Errors

Read the error status flags

pub fn clear_errors(&mut self, errors: Errors)

Clear error status flags

Setting a flag will clear the error. Clearing any flag will have no effect.

pub fn read_flags_errors(&self) -> Result<Flags, Error>

pub unsafe fn read_data(&mut self) -> u32

Read from the DATA register

Reading from the data register directly is unsafe, because it will clear the RXC flag, which could break assumptions made elsewhere in this module.

pub unsafe fn write_data(&mut self, data: u32)

Write to the DATA register

Writing to the data register directly is unsafe, because it will clear the DRE flag, which could break assumptions made elsewhere in this module.

pub fn disable(self) -> C

Disable the SPI peripheral and return the Config struct

impl<P, M> Spi<Config<P, M, DynLength>> where
P: ValidPads,
M: OpMode,
Config<P, M, DynLength>: ValidConfig,

pub fn get_dyn_length(&self) -> u8

Return the current transaction length

Read the LENGTH register to determine the current transaction length

pub unsafe fn set_dyn_length(&mut self, length: u8)

Set the transaction length

Write the LENGTH register to set the transaction length. Panics if the length is zero.

Safety

If you have sent any data at the current transaction length, you must wait for TXC before changing the length.

impl<C> Spi<C> where
C: ValidConfig,
C::Pads: Rx + NotTx,
C::OpMode: MasterMode,
C::Word: PrimInt,

pub unsafe fn reset_serial_read_state(&mut self)

Reset the internal state tracking serial Read transactions

See the implementation of Read for more details.