Trait atsamd_hal::sercom::v2::spi::AnyConfig

pub trait AnyConfig: Is<Type = SpecificConfig<Self>> {
    type Sercom: Sercom;
    type Pads: ValidPads<Sercom = Self::Sercom>;
    type OpMode: OpMode;
    type Length: Length<Word = Self::Word>;
    type Word: 'static;
}

Type class for all possible Config types

This trait uses the AnyKind trait pattern to create a type class for Config types. See the AnyKind documentation for more details on the pattern.

In addition to the normal, AnyKind associated types. This trait also copies the Sercom and Word types, to make it easier to apply bounds to these types at the next level of abstraction.

Associated Types

type Sercom: Sercom

type Pads: ValidPads<Sercom = Self::Sercom>

type OpMode: OpMode

type Length: Length<Word = Self::Word>

type Word: 'static

Loading content...