Trait atsamd_hal::sercom::v2::spi::AnySpi

pub trait AnySpi: Is<Type = SpecificSpi<Self>> {
    type Sercom: Sercom;
    type Pads: ValidPads<Sercom = Self::Sercom>;
    type OpMode: OpMode;
    type CharSize: CharSize<Word = Self::Word>;
    type Word: 'static;
    type Config: ValidConfig<Sercom = Self::Sercom, Pads = Self::Pads, OpMode = Self::OpMode, CharSize = Self::CharSize, Word = Self::Word>;
}

Type class for all possible Spi types

This trait uses the AnyKind trait pattern to create a type class for Spi types. See the AnyKind documentation for more details on the pattern.

In addition to the normal, AnyKind associated types. This trait also copies the Sercom, Pads, OpMode, CharSize and Word types, to make it easier to apply bounds to these types at the next level of abstraction.

Associated Types

type Sercom: Sercom

type Pads: ValidPads<Sercom = Self::Sercom>

type OpMode: OpMode

type CharSize: CharSize<Word = Self::Word>

type Word: 'static

type Config: ValidConfig<Sercom = Self::Sercom, Pads = Self::Pads, OpMode = Self::OpMode, CharSize = Self::CharSize, Word = Self::Word>

Loading content...