Trait atsamd_hal::sercom::v2::spi::CharSize

pub trait CharSize: Sealed {
    type Word: 'static;

    const BITS: u8;
    fn configure(sercom: &RegisterBlock) { ... }
}

Type-level enum representing the SPI character size

This trait acts as both a type-level enum, forming a type class for character sizes, as well as a type-level function mapping the corresponding word size.

The SPI character size affects the word size for the embedded HAL traits. Eight-bit transactions use a u8 word, while nine-bit transactions use a u16 word.

Associated Types

type Word: 'static

Word size for the character size

Loading content...
Associated Constants

const BITS: u8

Register it pattern for the corresponding CharSize

Loading content...
Provided methods

fn configure(sercom: &RegisterBlock)

Configure the LENGTH register and enable the LENGTH counter

Loading content...