Trait atsamd_hal::sercom::v2::spi::DipoDopo

pub trait DipoDopo {
    const DIPO: u8;
    const DOPO: u8;
    fn configure(sercom: &RegisterBlock) { ... }
}

Configure the DIPO and DOPO fields based on a set of Pads

According to the datasheet, the DIPO and DOPO values specify which SERCOM pads are used for various functions. Moreover, depending on which pads are actually in use, only certain combinations of these values make sense and are valid.

This trait is implemented for valid, four-tuple combinations of OptionalPadNums. Those implementations are then lifted to the corresponding Pads types.

To satisfy this trait, the combination of OptionalPadNums must specify a PadNum for CK and at least one of DI and DO. Furthermore, no two PadNums can conflict.

Associated Constants

const DIPO: u8

DIPO field value

const DOPO: u8

DIPO field value

Loading content...
Provided methods

fn configure(sercom: &RegisterBlock)

Configure the pad according to Self::DIPO and Self::DOPO

Loading content...