Trait atsamd_hal::sercom::v2::spi::OpMode

pub trait OpMode: Sealed {
    const MODE: MODE_A;
    const MSSEN: bool;
    fn configure(sercom: &RegisterBlock) { ... }
}

Type-level enum representing the SPI operating mode

See the documentation on type-level enums for a discussion of the pattern.

The available operating modes are Master, MasterHWSS and Slave. In Master mode, the SS signal must be handled by the user, so SS must be NoneT. In MasterHWSS mode, the hardware drives the SS line, so SomePad is required. In Slave mode, the SS pad is required as well, to indicate when data is valid.

Associated Constants

const MODE: MODE_A

Corresponding variant from the PAC enum

const MSSEN: bool

Bit indicating whether hardware SS control is enabled

Loading content...
Provided methods

fn configure(sercom: &RegisterBlock)

Configure the SPI operating mode

Loading content...