Trait atsamd_hal::sercom::v2::spi::TxOrRx

pub trait TxOrRx: ValidPads { }

Marker trait for a set of Pads that can transmit OR receive

To satisfy this trait, SCLK must always be SomePad and one or both of Data In and Data Out must also be SomePad.