Trait atsamd_hal::spi_common::CommonSpi

pub trait CommonSpi {
    fn spi(&self) -> &SPI;

    fn spi_mut(&mut self) -> &SPI;

    fn disable(&mut self) { ... }

    fn enable(&mut self) { ... }

    fn set_mode(&mut self, mode: Mode) { ... }

    fn freq<F: Into<Hertz>>(&self, src_clock_freq: Hertz) -> Hertz { ... }

    fn calculate_baud<F: Into<Hertz>>(freq: F, src_clock_freq: Hertz) -> u8 { ... }
}

Required methods

fn spi(&self) -> &SPI

Helper for accessing the spi member of the sercom instance

fn spi_mut(&mut self) -> &SPI

Helper for accessing the spi member of the sercom instance

Loading content...
Provided methods

fn disable(&mut self)

Disable the SPI

fn enable(&mut self)

Enable the SPI

fn set_mode(&mut self, mode: Mode)

Set the polarity (CPOL) and phase (CPHA) of the SPI

fn freq<F: Into<Hertz>>(&self, src_clock_freq: Hertz) -> Hertz

Method for calculating the output frequency given our baud settings.

for synchronous SERCOM peripherals, the calculation for the final frequency is f_baud = f_ref / (2 * (BAUD + 1)).

fn calculate_baud<F: Into<Hertz>>(freq: F, src_clock_freq: Hertz) -> u8

Helper for calculating our baudrate register

for synchronous SERCOM peripherals, the calculation for this register is BAUD = f_ref / (2 * f_baud) - 1.

Loading content...