Module atsamd_hal::thumbv6m

Modules

adc
Analogue-to-Digital Conversion

calibration
NVM Software Calibration Area Mapping

clock
Configuring the system clock sources. You will typically need to create an instance of GenericClockController before you can set up most of the peripherals on the atsamd21 device. The other types in this module are used to enforce at compile time that the peripherals have been correctly configured.

eic

pwm

timer
Working with timer counter hardware

usb
USB Device support

watchdog

Enums

ResetCause
ResetCause represents the reason the MCU was reset.

Functions

reset_cause
Returns the cause of the last reset.

serial_number
Returns the serial number of the chip as an array of bytes. The serial number is only guaranteed to be unique if all 16 bytes are used.

split_serial_number
Returns the serial number of the chip as 4 32-bit integers. The serial number is only guaranteed to be unique if all 128 bits are used.