Struct atsamd_hal::thumbv7em::adc::Adc

pub struct Adc<ADC> { /* fields omitted */ }

Adc encapsulates the device ADC

Implementations

impl Adc<ADC0>

pub fn adc0(
adc: ADC0,
mclk: &mut MCLK,
clocks: &mut GenericClockController,
gclk: GEN_A
) -> Self

pub fn samples(&mut self, samples: SampleRate)

Set the sample rate

pub fn reference(&mut self, reference: Reference)

Set the voltage reference

pub fn prescaler(&mut self, prescaler: Prescaler)

Set the prescaler for adjusting the clock relative to the system clock

pub fn resolution(&mut self, resolution: Resolution)

Set the input resolution

impl Adc<ADC1>

pub fn adc1(
adc: ADC1,
mclk: &mut MCLK,
clocks: &mut GenericClockController,
gclk: GEN_A
) -> Self

pub fn samples(&mut self, samples: SampleRate)

Set the sample rate

pub fn reference(&mut self, reference: Reference)

Set the voltage reference

pub fn prescaler(&mut self, prescaler: Prescaler)

Set the prescaler for adjusting the clock relative to the system clock

pub fn resolution(&mut self, resolution: Resolution)

Set the input resolution