Module atsamd_hal::thumbv7em::eic::pin

Structs

ExtInt0
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt1
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt2
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt3
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt4
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt5
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt6
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt7
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt8
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt9
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt10
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt11
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt12
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt13
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt14
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

ExtInt15
Represents a numbered external interrupt. The external interrupt is generic over any pin, only the EicPin implementations in this module make sense.

Traits

EicPin
The EicPin trait makes it more ergonomic to convert a gpio pin into an EIC pin. You should not implement this trait for yourself; only the implementations in the EIC module make sense.

ExternalInterrupt
ExternalInterrupt describes something with an external interrupt ID.

Type Definitions

ExternalInterruptID

Sense