Struct atsamd_hal::thumbv7em::timer::TimerCounter

pub struct TimerCounter<TC> { /* fields omitted */ }

A generic hardware timer counter. The counters are exposed in 16-bit mode only. The hardware allows configuring the 8-bit mode and pairing up some instances to run in 32-bit mode, but that functionality is not currently exposed by this hal implementation. TimerCounter implements both the Periodic and the CountDown embedded_hal timer traits. Before a hardware timer can be used, it must first have a clock configured.

Implementations

impl TimerCounter<TC2>

pub fn tc2_(clock: &Tc2Tc3Clock, tc: TC2, mclk: &mut MCLK) -> Self

Configure this timer counter instance. The clock is obtained from the GenericClockController instance and its frequency impacts the resolution and maximum range of the timeout values that can be passed to the start method. Note that some hardware timer instances share the same clock generator instance and thus will be clocked at the same rate.

impl TimerCounter<TC3>

pub fn tc3_(clock: &Tc2Tc3Clock, tc: TC3, mclk: &mut MCLK) -> Self

Configure this timer counter instance. The clock is obtained from the GenericClockController instance and its frequency impacts the resolution and maximum range of the timeout values that can be passed to the start method. Note that some hardware timer instances share the same clock generator instance and thus will be clocked at the same rate.