Module atsamd_hal::timer

Working with timer counter hardware

Structs

TimerCounter

A generic hardware timer counter. The counters are exposed in 16-bit mode only. The hardware allows configuring the 8-bit mode and pairing up some instances to run in 32-bit mode, but that functionality is not currently exposed by this hal implementation. TimerCounter implements both the Periodic and the CountDown embedded_hal timer traits. Before a hardware timer can be used, it must first have a clock configured.

Traits

Count16

This is a helper trait to make it easier to make most of the TimerCounter impl generic. It doesn’t make too much sense to to try to implement this trait outside of this module.

Type Definitions

TimerCounter3

TimerCounter4

TimerCounter5