Crate atsamd_hal

Re-exports

pub extern crate embedded_hal as hal;

pub use paste;

pub use atsamd11c as target_device;

pub use self::common::*;

pub use self::samd11::*;

Modules

common

samd11

typelevel
Module supporting type-level programming

Macros

dbgprint

define_pins
This macro is a helper for defining a Pins type in a board support crate. This type is used to provide more meaningful aliases for the various GPIO pins for a given board.