atsamd_hal

Module adc

Analogue-to-Digital Conversion

Structs

Adc

Adc encapsulates the device ADC

Enums

Gain

The gain level Gain Factor Selection

Prescaler

Clock frequency relative to the system clock Prescaler Configuration

Reference

Reference voltage (or its source) Reference Selection

Resolution

Reading resolution in bits

SampleRate

Samples per reading Number of Samples to be Collected