atsamd_hal::aes

Struct Aes

pub struct Aes { /* private fields */ }

AES Peripheral

Encapsulates the PAC which acts as a token and provides an interface to the AES hardware

Implementations

impl Aes

pub fn new(aes: Aes) -> Self

Create the interface for the AES peripheral

Don’t forget to enable the APB bus for AES

AHB bus is on by default at reset

Clock::v1 mclk.apbcmask.modify(|_, w| w.aes_().set_bit());

Clock::v2 tokens.apbs.aes.enable();

pub fn enable(&mut self)

Enable the AES peripheral

pub fn disable(&mut self)

Disable the AES peripheral

pub fn swrst(&self)

Reset the AES controller

pub fn destroy(self) -> Aes

Destroy the AES peripheral and return the underlying AES register

pub fn set_aesmode(self, mode: Aesmodeselect)

Set AES Mode

pub fn set_cfbs(self, blocksize: Cfbsselect)

Set Cipher Feedback Block Size (CFBS)

pub fn set_keysize(self, keysize: Keysizeselect)

Set Encryption Key Size

pub fn set_cipher(self, mode: Cipherselect)

Set Cipher Mode

pub fn set_startmode(self, mode: Startmodeselect)

Set Start Mode

pub fn set_lod(self, mode: Lodselect)

Set Last Output Data (LOD) Mode

pub fn set_keygen(self, keygen_start: bool)

Start Last Key Generation

Compute last NK words of expanded key

pub fn set_xorkey(self, mode: Xorkeyselect)

XOR Key Generation

The user keyword gets XORed with the previous keyword register content

pub fn set_ctype(self, countermeasures: Ctype)

Counter Measure Type

pub fn start(&self)

Start Encryption/Decryption

pub fn newmsg(&self)

New Message Used in cipher block chaining (CBC), cipher feedback (CFB) and output feedback (OFB), counter (CTR) modes to indicate that the hardware should use Initialization Vector for encrypting the first block of a message.

pub fn eom(&self)

End of Message (GCM mode only)

This triggers generation of final GHASH value for the message

pub fn gfmul(&self)

GF Multiplication (GCM mode only)

This triggers GF multiplication of data buffer content and hashkey register content.

pub fn disable_enccmp(&self)

Disable Encryption Complete Interrupt

pub fn disable_gfmcmp(&self)

Disable GF Multiplication Complete Interrupt

pub fn enable_enccmp(&self)

Enable Encryption Complete Interrupt

pub fn enable_gfmcmp(&self)

Enable GF Multiplication Complete Interrupt

pub fn clear_enccmp(&self)

Clear Encryption Complete Interrupt

Also automatically cleared if

Manual encryption/decryption occurs (via CTRLB.START)

Reading from GHASHx register

pub fn clear_gfmcmp(&self)

Clear GF Multiplication Complete Interrupt

Also automatically cleared if

Manual encryption/decryption occurs (via CTRLB.START)

Reading from data register INDATAx when LOD = 0

Writing into the data register INDATAx when LOD = 1

Reading from Hash Key register (HASHKEYx)

pub fn read_enccmp(&self) -> bool

Read Encryption Complete Interrupt

pub fn read_gfmcmp(&self) -> bool

Read GF Multiplication Complete Interrupt

pub fn set_databufptr(&self, dataptr: u8)

Set the Data Buffer Pointer

pub fn set_debug(&self, run_during_debug: bool)

Control if AES is active during debugging

Enable protected, only possible to change when AES is disabled

pub fn set_keyword<const N: usize>(&self, keyword: &[u8; N])

Set keyword / cryptographic key

Consists of four/six/eight Key Word registers for setting the cryptographic key

pub fn set_data(&self, data: u32)

Set indata

pub fn get_data(&self) -> u32

Read indata

pub fn set_initialization_vector(&self, iv: [u32; 4])

Set initialization vector

The four 32-bit Initialization Vector registers INTVECTVn set the 128-bit Initialization Vector data block that is used by some modes of operation as an additional initial input. INTVECTV0.INTVECTV corresponds to the first word of the Initialization Vector, INTVECTV3.INTVECTV to the last one.

These registers are write-only to prevent the Initialization Vector from being read by another application. For CBC, OFB, and CFB modes, the Initialization Vector corresponds to the initialization vector. For CTR mode, it corresponds to the counter value.

pub fn set_hashkey(&self, hashkey: [u32; 4])

Set GCM hash key value (GCM mode only)

pub fn get_hashkey(&self) -> [u32; 4]

pub fn set_ghash(&self, ghash: [u32; 4])

Set Galois Hash value (GCM mode only)

Writing a new key to keyword clears GHASH

pub fn get_ghash(&self) -> [u32; 4]

Read Galois Hash value (GCM mode only)

pub fn set_ciplen(&self, ciplen: u32)

Set cipher length (GCM mode only)

pub fn get_ciplen(&self) -> u32

Read Cipher Length (GCM mode only)

pub fn set_randseed(&self, seed: u32)

Set a new seed for the AES random number generator