atsamd_hal

Module clock

Configuring the system clock sources.

You will typically need to create an instance of GenericClockController before you can set up most of the peripherals on the atsamd21 device. The other types in this module are used to enforce at compile time that the peripherals have been correctly configured.

Structs

AcAnaClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

AcDigClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

AdcClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

DacClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

EicClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys0Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys1Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys2Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys3Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys4Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys5Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys6Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys7Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys8Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys9Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys10Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys11Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

GClock

Represents a configured clock generator.

GenericClockController

GenericClockController encapsulates the Gclk hardware.

I2S0Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

I2S1Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

RtcClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom0CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom1CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom2CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom3CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom4CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom5CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tc4Tc5Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tc6Tc7Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tcc0Tcc1Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tcc2Tc3Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

UsbClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

WdtClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Constants

OSC8M_FREQ

The frequency of the 8 Mhz source.

OSC32K_FREQ

The frequency of the 32Khz source.

OSC48M_FREQ

The frequency of the 48Mhz source.

Functions

enable_external_32kosc

Turn on the external 32hkz oscillator

enable_internal_32kosc

Turn on the internal 32hkz oscillator

Type Aliases

ClockGenId

ClockId

ClockSource