atsamd_hal::clock
Module v1

Clocking API v1

Configuring the system clock sources. You will typically need to create an instance of GenericClockController before you can set up most of the peripherals on the atsamd51 device. The other types in this module are used to enforce at compile time that the peripherals have been correctly configured.

Structs

AcClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Adc0Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Adc1Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

CclClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Cm4TraceClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

DacClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

EicClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys0Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys1Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys2Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys3Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys4Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys5Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys6Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys7Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys8Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys9Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys10Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Evsys11Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

FreqmMsrClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

FreqmRefClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

GClock

Represents a configured clock generator.

GenericClockController

GenericClockController encapsulates the Gclk hardware.

I2S0Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

I2S1Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

PdecClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sdhc0Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sdhc1Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom0CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom1CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom2CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom3CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom4CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom5CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom6CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Sercom7CoreClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tc0Tc1Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tc2Tc3Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tc4Tc5Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tc6Tc7Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tcc0Tcc1Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tcc2Tcc3Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Tcc4Clock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

UsbClock

A typed token that indicates that the clock for the peripheral(s) with the matching name has been configured.

Enums

ClockId

Constants

OSC32K_FREQ

The frequency of the 32Khz source.

OSC48M_FREQ

The frequency of the 48Mhz source.

OSC120M_FREQ

The frequency of the 120Mhz source.

Type Aliases

ClockGenId

ClockSource