atsamd_hal::clock::v2::ahb

Trait AhbId

pub trait AhbId: Sealed {
    const DYN: DynAhbId;
}

Type-level enum identifying one of the possible AHB clocks

The types implementing this trait are type-level variants of AhbId, and they identify one of the possible AHB clocks, which can vary by chip. Each type corresponds to a specific bit in the AHBMASK register.

AhbId is the type-level equivalent of DynAhbId. See the documentation on type-level programming and specifically type-level enums for more details.

Required Associated Constants

const DYN: DynAhbId

Corresponding DynAhbId

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.