atsamd_hal::clock::v2::apb

Enum DynApbId

#[repr(u8)]
pub enum DynApbId {

    Pac = 0,
    Pm = 1,
    Mclk = 2,
    RstC = 3,
    OscCtrl = 4,
    Osc32kCtrl = 5,
    SupC = 6,
    Gclk = 7,
    Wdt = 8,
    Rtc = 9,
    Eic = 10,
    FreqM = 11,
    Sercom0 = 12,
    Sercom1 = 13,
    Tc0 = 14,
    Tc1 = 15,
    Usb = 16,
    Dsu = 17,
    NvmCtrl = 18,
    Port = 19,
    EvSys = 20,
    Sercom2 = 21,
    Sercom3 = 22,
    Tcc0 = 23,
    Tcc1 = 24,
    Tc2 = 25,
    Tc3 = 26,
    RamEcc = 27,
    Tcc2 = 28,
    PDec = 29,
    Ac = 30,
    Aes = 31,
    Trng = 32,
    Icm = 33,
    Qspi = 34,
    Ccl = 35,
    Sercom4 = 36,
    Sercom5 = 37,
    Adc0 = 38,
    Adc1 = 39,
    Dac = 40,
    Pcc = 41,

}

Value-level enum identifying a single APB clock

Each variant of this enum corresponds to a specific bit in one of the four APBXMASK registers and identifies one of many possible APB clocks, which can vary by chip.

DynApbId is the value-level equivalent of ApbId.

Variants

Pac = 0

Pm = 1

Mclk = 2

RstC = 3

OscCtrl = 4

Osc32kCtrl = 5

SupC = 6

Gclk = 7

Wdt = 8

Rtc = 9

Eic = 10

FreqM = 11

Sercom0 = 12

Sercom1 = 13

Tc0 = 14

Tc1 = 15

Usb = 16

Dsu = 17

NvmCtrl = 18

Port = 19

EvSys = 20

Sercom2 = 21

Sercom3 = 22

Tcc0 = 23

Tcc1 = 24

Tc2 = 25

Tc3 = 26

RamEcc = 27

Tcc2 = 28

PDec = 29

Ac = 30

Aes = 31

Trng = 32

Icm = 33

Qspi = 34

Ccl = 35

Sercom4 = 36

Sercom5 = 37

Adc0 = 38

Adc1 = 39

Dac = 40

Pcc = 41