atsamd_hal::clock::v2::dfll

Trait Mode

pub trait Mode {
    const DYN: DynMode;
}

Type-level enum identifying the Dfll control loop mode

The types implementing this trait, i.e. OpenLoop, FromPclk and FromUsb, are type-level variants of Mode, and they determine whether the DFLL operates in closed-loop mode, and if so, which Reference clock to use.

Mode is the type-level equivalent of DynMode. See the documentation on type-level programming and specifically type-level enums for more details.

Required Associated Constants

const DYN: DynMode

Corresponding variant of DynMode

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.