atsamd_hal::clock::v2::dpll

Type Alias EnabledDpll

pub type EnabledDpll<D, I, N = U0> = Enabled<Dpll<D, I>, N>;

An Enabled Dpll

As described in the clock module documentation, the Enabled wrapper implements compile-time clock tree safety by tracking the number of consumer clocks and restricting access to the underlying Dpll to prevent modification while in use.

As with Enabled, the default value for N is U0; if left unspecified, the counter is assumed to be zero.

Aliased Type

struct EnabledDpll<D, I, N = U0>(/* private fields */);

Implementations

impl<D, I> EnabledDpll<D, I>
where D: DpllId, I: DpllSourceId,

pub fn disable(self) -> Dpll<D, I>

Disable the Dpll

This method is only implemented for N = U0, which means the clock can only be disabled when no other clocks consume this Dpll.

impl<D, I, N> EnabledDpll<D, I, N>
where D: DpllId, I: DpllSourceId,

pub fn is_locked(&self) -> bool

Test whether the Dpll is locked

pub fn is_ready(&self) -> bool

Test whether the Dpll is ready