atsamd_hal::clock::v2::gclk
Trait GclkIo

pub trait GclkIo: PinId {
    type GclkId: GclkId;
}

Trait mapping each PinId to its corresponding GclkId when used as a Gclk input or output

If a given PinId can be used as a Gclk input or output, it can only be used with one specific GclkId. This trait provides a mapping from such a PinId to its corresponding GclkId.

Required Associated Types

type GclkId: GclkId

Corresponding GclkId for this PinId

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.