atsamd_hal::clock::v2::gclk

Type Alias EnabledGclk0

pub type EnabledGclk0<I, N = U1> = EnabledGclk<Gclk0Id, I, N>;

Type alias for the corresponding EnabledGclk

As mentioned in the module-level documentation, Gclk0 is special, because it provides the processor main clock. We represent this by permanently Incrementing the counter for EnabledGclk0, which prevents it from ever being disabled. Thus, the default value for N is U1 instead of U0. Accordingly, we also provide a few special methods on EnabledGclk0 to configure the Gclk while it is actively running.

Aliased Type

struct EnabledGclk0<I, N = U1>(/* private fields */);

Implementations

impl<I: GclkSourceId> EnabledGclk0<I, U1>

Special methods for an Enabled Gclk0

Gclk0 is special, because it drives the processor’s main clock, which can never be disabled. As discussed in the module-level documentation, this fact is represented by permanently Incrementing the counter for EnabledGclk0. Thus, the minimum value for N is U1 and EnabledGclk0 can never be disabled.

These methods represent actions that can be taken when N = U1, i.e. the Enabled counter is at its minimum value. This is the only time it’s safe to change the Gclk0 Source or change its GclkDivider value.

pub fn swap_sources<O, N>( self, old: O, new: N, ) -> (EnabledGclk0<N::Id, U1>, O::Dec, N::Inc)
where O: Source<Id = I> + Decrement, N: Source + Increment, N::Id: NotGclkIo,

Swap Gclk0 from one clock Source to another

Gclk0 will remain fully enabled during the swap.

pub fn swap_pins<P>( self, pin: P, freq: impl Into<Hertz>, ) -> (EnabledGclk0<P::Id, U1>, Pin<I, AlternateM>)
where I: Gclk0Io, P: AnyPin, P::Id: Gclk0Io,

Swap Gclk0 from one GclkIo Pin to another

Gclk0 will remain fully enabled during the swap.

pub fn swap_source_for_pin<S, P>( self, source: S, pin: P, freq: impl Into<Hertz>, ) -> (EnabledGclk0<P::Id, U1>, S::Dec)
where S: Source<Id = I> + Decrement, P: AnyPin, P::Id: Gclk0Io,

Swap Gclk0 from a clock Source to a GclkIo Pin

Gclk0 will remain fully enabled during the swap.

pub fn swap_pin_for_source<S>( self, source: S, ) -> (EnabledGclk0<S::Id, U1>, Pin<I, AlternateM>, S::Inc)
where I: Gclk0Io, S: Source + Increment, S::Id: NotGclkIo,

Swap Gclk0 from a GclkIo Pin to a clock Source

Gclk0 will remain fully enabled during the swap.

pub fn div(&mut self, div: GclkDiv8)

Set the GclkDivider value for Gclk0

See Gclk::div documentation for more details.

pub fn improve_duty_cycle(&mut self, flag: bool)

Output a 50-50 duty cycle clock when using an odd GclkDivider

pub fn freq(&self) -> Hertz

Return the Gclk0 frequency

See Gclk::freq documentation for more details.

pub fn output_off_value(&mut self, high: bool)

Set the state of GclkOut pins when GclkIo output is disabled

See Gclk::output_off_value documentation for more details.