atsamd_hal::clock::v2::rtcosc
Trait RtcSourceId

pub trait RtcSourceId {
    const DYN: DynRtcSourceId;
    const FREQ: Hertz;
}

Type-level enum for the RTC oscillator clock source

The RTC can be sourced from any of OscUlp1k, OscUlp32k, Xosc1k or Xosc32k.

See the type-level enum documentation for more details on the pattern.

Type-level enum of possible clock Sources for the RtcOsc

The types implementing this trait are type-level variants of RtcSourceId, and they identify one of four possible clock Sources for the RtcOsc. All implementers of this trait are Id types, which are described in more detail in the clock module documentation.

RtcSourceId is the type-level equivalent of DynRtcSourceId. See the documentation on type-level programming and specifically type-level enums for more details.

Required Associated Constants

const DYN: DynRtcSourceId

Corresponding DynRtcSourceId

const FREQ: Hertz

Clock source frequency, either 1 kHz or 32 kHz

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.