atsamd_hal::clock::v2
Module types

Module defining or exporting peripheral types for the [‘ahb’], [‘apb’] and [‘pclk’] modules

The ahb, apb and pclk modules each define structs that are generic over a type parameter representing a peripheral. Some peripheral modules already define suitable types for this purpose. For example, sercom defines the Sercom0, Sercom1, etc. types. But other peripherals are either not yet implemented in the HAL or do not define a suitable type. This module defines a type for such peripherals. If/when a suitable type is added for a given peripheral, the type defined here should be deprecated or removed.

Re-exports

pub use crate::sercom::Sercom0;

pub use crate::sercom::Sercom1;

pub use crate::sercom::Sercom2;

pub use crate::sercom::Sercom3;

pub use crate::sercom::Sercom4;

pub use crate::sercom::Sercom5;

pub use crate::sercom::Sercom6;

pub use crate::sercom::Sercom7;

Enums

Ac

Marker type representing the corresponding peripheral

Adc0

Marker type representing the corresponding peripheral

Adc1

Marker type representing the corresponding peripheral

Aes

Marker type representing the corresponding peripheral

CM4Trace

Marker type representing the corresponding peripheral

Ccl

Marker type representing the corresponding peripheral

Cmcc

Marker type representing the corresponding peripheral

Dac

Marker type representing the corresponding peripheral

Dmac

Marker type representing the corresponding peripheral

Dsu

Marker type representing the corresponding peripheral

Eic

Marker type representing the corresponding peripheral

EvSys

Marker type representing the corresponding peripheral

EvSys0

Marker type representing the corresponding peripheral

EvSys1

Marker type representing the corresponding peripheral

EvSys2

Marker type representing the corresponding peripheral

EvSys3

Marker type representing the corresponding peripheral

EvSys4

Marker type representing the corresponding peripheral

EvSys5

Marker type representing the corresponding peripheral

EvSys6

Marker type representing the corresponding peripheral

EvSys7

Marker type representing the corresponding peripheral

EvSys8

Marker type representing the corresponding peripheral

EvSys9

Marker type representing the corresponding peripheral

EvSys10

Marker type representing the corresponding peripheral

EvSys11

Marker type representing the corresponding peripheral

FreqM

Marker type representing the corresponding peripheral

FreqMMeasure

Marker type representing the corresponding peripheral

FreqMReference

Marker type representing the corresponding peripheral

Gclk

Marker type representing the corresponding peripheral

Hpb0

Marker type representing the corresponding peripheral

Hpb1

Marker type representing the corresponding peripheral

Hpb2

Marker type representing the corresponding peripheral

Hpb3

Marker type representing the corresponding peripheral

I2S

Marker type representing the corresponding peripheral

I2S0

Marker type representing the corresponding peripheral

I2S1

Marker type representing the corresponding peripheral

Icm

Marker type representing the corresponding peripheral

Mclk

Marker type representing the corresponding peripheral

NvmCtrl

Marker type representing the corresponding peripheral

NvmCtrlCache

Marker type representing the corresponding peripheral

NvmCtrlSmeeProm

Marker type representing the corresponding peripheral

Osc32kCtrl

Marker type representing the corresponding peripheral

OscCtrl

Marker type representing the corresponding peripheral

PDec

Marker type representing the corresponding peripheral

Pac

Marker type representing the corresponding peripheral

Pcc

Marker type representing the corresponding peripheral

Pm

Marker type representing the corresponding peripheral

Port

Marker type representing the corresponding peripheral

Pukcc

Marker type representing the corresponding peripheral

Qspi

Marker type representing the corresponding peripheral

Qspi2x

Marker type representing the corresponding peripheral

RamEcc

Marker type representing the corresponding peripheral

RstC

Marker type representing the corresponding peripheral

Rtc

Marker type representing the corresponding peripheral

Sdhc0

Marker type representing the corresponding peripheral

Sdhc1

Marker type representing the corresponding peripheral

SlowClk

Marker type representing the corresponding peripheral

SupC

Marker type representing the corresponding peripheral

Tc0

Marker type representing the corresponding peripheral

Tc0Tc1

Marker type representing the corresponding peripheral

Tc1

Marker type representing the corresponding peripheral

Tc2

Marker type representing the corresponding peripheral

Tc3

Marker type representing the corresponding peripheral

Tc4

Marker type representing the corresponding peripheral

Tc5

Marker type representing the corresponding peripheral

Tc6

Marker type representing the corresponding peripheral

Tc7

Marker type representing the corresponding peripheral

Tc2Tc3

Marker type representing the corresponding peripheral

Tc4Tc5

Marker type representing the corresponding peripheral

Tc6Tc7

Marker type representing the corresponding peripheral

Tcc0

Marker type representing the corresponding peripheral

Tcc0Tcc1

Marker type representing the corresponding peripheral

Tcc1

Marker type representing the corresponding peripheral

Tcc2

Marker type representing the corresponding peripheral

Tcc3

Marker type representing the corresponding peripheral

Tcc4

Marker type representing the corresponding peripheral

Tcc2Tcc3

Marker type representing the corresponding peripheral

Trng

Marker type representing the corresponding peripheral

Usb

Marker type representing the corresponding peripheral

Wdt

Marker type representing the corresponding peripheral