atsamd_hal::clock::v2::xosc

Type Alias EnabledXosc

pub type EnabledXosc<X, M, N = U0> = Enabled<Xosc<X, M>, N>;

An Enabled Xosc

As described in the clock module documentation, the Enabled wrapper implements compile-time clock tree safety by tracking the number of consumer clocks and restricting access to the underlying Xosc to prevent modification while in use.

As with Enabled, the default value for N is U0; if left unspecified, the counter is assumed to be zero.

Aliased Type

struct EnabledXosc<X, M, N = U0>(/* private fields */);

Implementations

impl<X, M> EnabledXosc<X, M>
where X: XoscId, M: Mode,

pub fn disable(self) -> Xosc<X, M>

Disable the Xosc

This method is only implemented for N = U0, which means the clock can only be disabled when no other clocks consume this Xosc.

impl<X, M, N> EnabledXosc<X, M, N>
where X: XoscId, M: Mode,

pub fn is_ready(&self) -> bool

Check whether the Xosc is stable and ready to be used as a clock Source

pub fn enable_failure_detection<S>( &mut self, dfll: S, div: SafeClockDiv, ) -> S::Inc
where S: Source<Id = DfllId> + Increment,

Enable continuous monitoring of the Xosc for clock failure

Failure detection will continuously monitor the Xosc to verify it is still running. In the event of a failure, the Xosc output will be switched to the “safe clock”.

The safe clock is derived from the DFLL, which runs at 48 MHz. The XOSC peripheral provides a prescaler to divide down the 48 MHz DFLL to better match the clock it replaces. The prescaler division factor can be any power of two, 2^N, with N in the range 0..16.

For example, if the Xosc input frequency is 16 MHz, a reasonable divider would be 4, becuase the safe clock frequency would be 12 MHz, which is closest to 16 MHz.

Note that clock failure is triggered when four safe clock periods pass without seeing a rising & falling edge pair on the XOSC clock. Once failure is detected, the corresponding bit in the STATUS register will go high and an interrupt will be triggered.

If the external clock can be fixed, the Xosc can be switched back to it using EnabledXosc::switch_back.

Because the safe clock makes use of the DFLL, the Xosc must register as a consumer of the EnabledDfll and Increment its counter.

pub fn has_failed(&self) -> bool

Check whether the Xosc has triggered clock failure detection

Failure detection must be enabled for this to return true. Failure is triggered when four safe clock periods pass without seeing a rising & falling edge pair on the XOSC clock.

See EnabledXosc::enable_failure_detection for more details.

pub fn is_switched(&self) -> bool

Check whether the Xosc has been switched to the safe clock

Returns true if the Xosc has been switched to the safe clock. This will only occur if clock failure detection is enabled.

pub fn switch_back(&mut self)

Attempt to switch from the safe clock back to the external clock

This function will set the switch back bit (SWBEN) in the XOSCCTRL register. Once the hardware has successfully switched back, this bit will be automatically cleared.

Users can check whether switching back was successful by checking the STATUS register with EnabledXosc::is_switched.

pub fn disable_failure_detection<S>(&mut self, dfll: S) -> S::Dec
where S: Source<Id = DfllId> + Decrement,

Disable continuous monitoring of the Xosc for clock failure

Once failure monitoring is disabled, the DFLL is no longer used as the safe clock, so the EnabledDfll counter can be Decremented.