Trait atsamd_hal::common::sercom::v1::spi::DipoDopo

pub trait DipoDopo {
    fn dipo_dopo(&self) -> (u8, u8);
}

The DipoDopo trait defines a way to get the data in and data out pin out values for a given SPIMasterXPadout configuration. You should not implement this trait for yourself; only the implementations in the sercom module make sense.

Required methods

fn dipo_dopo(&self) -> (u8, u8)

Loading content...