atsamd_hal::eic

Struct ExtInt

pub struct ExtInt<P, Id, F = NoneT>
where
    P: EicPin,
    Id: ChId,
{ /* private fields */ }

A numbered external interrupt, which can be used to sense state changes on its pin.

Implementations

impl<P, Id> ExtInt<P, Id, NoneT>
where P: EicPin, Id: ChId,

pub fn into_future<I>(self, _irq: I) -> ExtInt<P, Id, EicFuture>
where I: Binding<P::InterruptSource, InterruptHandler>, InterruptHandler: Handler<P::InterruptSource>,

Turn an EIC pin into a pin usable as a Future. The correct interrupt source is needed.

impl<P, Id> ExtInt<P, Id, EicFuture>
where P: EicPin, Self: InputPin<Error = Infallible>, Id: ChId,

pub async fn wait(&mut self, sense: Sense)

impl<P, Id, F> ExtInt<P, Id, F>
where P: EicPin, Id: ChId,

pub fn enable_event(&mut self)

pub fn enable_interrupt(&mut self)

pub fn disable_interrupt(&mut self)

pub fn is_interrupt(&mut self) -> bool

pub fn state(&mut self) -> bool

pub fn clear_interrupt(&mut self)

pub fn sense(&mut self, sense: Sense)

pub fn filter(&mut self, filter: bool)

pub fn debounce(&mut self)

Enable debouncing for this pin, with a configuration appropriate for debouncing physical buttons.

impl<P, Id, F> ExtInt<P, Id, F>
where P: EicPin, Id: ChId,

pub fn free(self) -> (Pin<P::Id, P::Mode>, Channel<Id, F>)

Release the underlying resources: Pin and Channel.