atsamd_hal::gpio::pin

Trait PinMode

pub trait PinMode: Sealed {
    const DYN: DynPinMode;
}

Type-level enum representing pin modes

The valid options are Disabled, Input, Output and Alternate. See the type-level enum documentation for more details on the pattern.

Required Associated Constants

const DYN: DynPinMode

Corresponding DynPinMode

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.