atsamd_hal::nvm

Struct Nvm

pub struct Nvm { /* private fields */ }

Non-volatile memory controller

Implementations

impl Nvm

pub const USERPAGE_ADDR: *const [u8; 512] = {0x804000 as *const [u8; 512]}

Pointer to the userpage region of the flash memory

Note: Never call core::ptr::read_volatile on this pointer, yields very poor codegen (around 9 KiB bytes of code)

pub fn new(nvm: Nvmctrl) -> Self

Create a new NVM controller or handle failure from DSU

pub unsafe fn registers(&self) -> &Nvmctrl

Raw access to the registers.

Safety

The abstraction assumes that it has exclusive ownership of the registers. Direct access can break such assumptions.

pub unsafe fn bank_swap(&mut self) -> !

Swap the flash banks. The processor will be reset, after which the inactive bank will become the active bank.

Safety

Ensure there is a working, memory safe program in place in the inactive bank before calling.

pub fn power_reduction_mode(&mut self, prm: Prmselect)

Set the power reduction mode

pub fn is_boot_protected(&self) -> bool

Check if the flash is boot protected

pub fn first_bank(&self) -> PhysicalBank

Get first bank

pub fn read_userpage(&self) -> Userpage

Read the user page from the flash memory

pub unsafe fn modify_userpage( &mut self, f: impl FnOnce(&mut Userpage), ) -> Result<UserpageStatus>

Modify the NVM User Page (aka User Row/UROW)

User is expected to provide a closure that modifies the user page according to the user’s needs.

This method will read the current user page, call the closure on it, erase the page in the flash memory and write it back again.

Erasure and flashing is skipped if the userpage stays the same after calling the closure on it.

Safety

Even though factory calibration settings are not modifiable via setters they can be still mutated via raw access to the userpage.0 field.

Power loss between the erase and the write will result in data loss.

Thus, users are advised to backup factory calibration settings before mutating the user page!

If these settings are erased, device might stop behaving correctly!

pub fn calibration_area(&self) -> CalibrationArea

Read the calibration area

pub fn temperatures_calibration_area(&self) -> TemperaturesCalibrationArea

Read the calibration area for temperatures

pub fn enable_security_bit(&mut self) -> Result<()>

Enable security bit

It locks the chip from external access for code security. Consult the datasheet for more details.

In order to disable it, chip erase command must be issued through the debugger.

pub unsafe fn enable_chip_erase_lock(&mut self) -> Result<()>

Enable the chip erase lock

It disables the chip erase capability.

Safety

Together with Self::enable_security_bit, it completely locks the MCU down from any external interaction via debugger, thus effectively bricking the device. Flashed firmware must provide a way to execute Self::disable_chip_erase_lock method in order to enable the debugger access again.

pub fn disable_chip_erase_lock(&mut self) -> Result<()>

Disable the chip erase lock

It enables the chip erase capability through the debugger.

pub fn boot_protection(&mut self, protect: bool) -> Result<()>

Enable/disable boot protection

Userpage’s NVM BOOT field defines a memory region that is supposed to be protected. Nvmctrl.STATUS.BOOTPROT is a readonly HW register populated on reset with a value from a userpage. By default, 0

pub fn region_lock(&mut self, mask: u32) -> Result<()>

Enable/disable region lock

Flash memory is split into 32 regions. The 32 bits of the mask determine if each region should be locked (if its bit is 0) and prevent writing and erasing pages, or unlocked (if its bit is 1) and allow writing and erasing pages.

Less significant bits represent lower addresses, more significant bits represent higher addresses.

For example mask 0xFFFF_0000 implies that active bank should be locked while inactive bank should be unlocked.

pub unsafe fn write_flash_from_slice( &mut self, destination: *mut u32, source_slice: &[u32], write_granularity: WriteGranularity, ) -> Result<()>

Write to the main address space flash memory from a slice

This call will fail if area that is being written to is

outside of the main address space flash area

write protected (BOOTPROT)

overlapping with SmartEEPROM flash region

destination has to be 4 bytes aligned.

Safety

Writes to the main address space flash area containing currently executed application are unsound.

pub unsafe fn write_flash( &mut self, destination: *mut u32, source: *const u32, words: u32, write_granularity: WriteGranularity, ) -> Result<()>

Write to the main address space flash memory

This call will fail if area that is being written to is

outside of the main address space flash area

write protected (BOOTPROT)

overlapping with SmartEEPROM flash region

destination has to be 4 bytes aligned. source has to be 4 bytes aligned.

Safety

Writes to the main address space flash area containing currently executed application are unsound.

pub unsafe fn erase_flash( &mut self, address: *mut u32, blocks: u32, ) -> Result<()>

Erase the portion of the main address space flash memory

Erase granularity is expressed in blocks (16 pages == 8192 bytes)

This call will fail if area that is being erased is

outside of the main address space flash area

write protected (BOOTPROT)

overlapping with SmartEEPROM flash region

Safety

Erasure of the main address space flash area containing currently executed application is unsound.

pub fn smart_eeprom(&mut self) -> Result<'_>

Retrieve SmartEEPROM