atsamd_hal::pukcc::c_abi

Trait CryptoRamSlice

pub trait CryptoRamSlice: Sealed {
    // Required method
    unsafe fn pukcc_base(&self) -> nu1;
}

Trait implemented for all &[u8] slices in order to provide a normalized way of downcasting pointers in a form accepted by PUKCC ABI

Required Methods

unsafe fn pukcc_base(&self) -> nu1

Take a slice base-pointer and mask out a halfword according to the expected format of CryptoRAM pointers in PukclParams-related structs.

Safety

Might break aliasing rules if more than one instance is used.

Only to be used exclusively for low-level access, never together with the high level API.