Struct atsamd_hal::rtc::Rtc

pub struct Rtc<Mode: RtcMode> { /* private fields */ }

Rtc represents the RTC peripheral for either clock/calendar or timer mode.

Implementations

impl<Mode: RtcMode> Rtc<Mode>

pub fn into_count32_mode(self) -> Rtc<Count32Mode>

Reonfigures the peripheral for 32bit counter mode.

pub fn into_clock_mode(self) -> Rtc<ClockMode>

Reconfigures the peripheral for clock/calendar mode. Requires the source clock to be running at 1024 Hz.

pub fn free(self) -> RTC

Releases the RTC resource

impl Rtc<Count32Mode>

pub fn count32_mode(rtc: RTC, rtc_clock_freq: Hertz, pm: &mut PM) -> Self

Configures the RTC in 32-bit counter mode with no prescaler (default state after reset) and the counter initialized to zero.

pub fn count32(&self) -> u32

Returns the internal counter value.

pub fn set_count32(&mut self, count: u32)

Sets the internal counter value.

pub fn reset_and_compute_prescaler<T: Into<<Self as CountDown>::Time>>(
&mut self,
timeout: T
) -> &Self

This resets the internal counter and sets the prescaler to match the provided timeout. You should configure the prescaler using the longest timeout you plan to measure.

impl Rtc<ClockMode>

pub fn clock_mode(rtc: RTC, rtc_clock_freq: Hertz, pm: &mut PM) -> Self

pub fn current_time(&self) -> Datetime

Returns the current clock/calendar value.

pub fn set_time(&mut self, time: Datetime)

Updates the current clock/calendar value.