Module atsamd_hal::sercom

Configure the SERCOM peripherals

The SERCOM module is used to configure the SERCOM peripherals as USART, SPI or I2C interfaces.

Undocumented features

The ATSAMx5x chips contain certain features that aren’t documented in the datasheet. These features are implemented in the HAL based on experimentation with certain boards which have verifiably demonstrated that those features work as intended.

UndocIoSet1: Implement an undocumented IoSet for PA16, PA17, PB22 & PB23 configured for Sercom1. The pygamer & feather_m4 use this combination, but it is not listed as valid in the datasheet.

UndocIoSet2: Implement an undocumented IoSet for PA00, PA01, PB22 & PB23 configured for Sercom1. The itsybitsy_m4 uses this combination, but it is not listed as valid in the datasheet.

PB02 is I2C-capable according to metro_m4. As such, PB02 implements IsI2cPad.

PB03 is I2C-capable according to metro_m4. As such, PB03 implements IsI2cPad.

Re-exports

pub use pad::*;

Modules

i2c

Use the SERCOM peripheral for I2C communications

pad

Type-level tools to configure SERCOM pads

spi

Use a SERCOM peripheral for SPI transactions

spi_future

A Future-like interface for SPI transactions

uart

Use the SERCOM peripheral for UART communications

Traits

Sercom

Type-level enum representing a Serial Communication Interface (SERCOM)

Type Definitions

Sercom0

Type alias for the corresponding SERCOM instance

Sercom1

Type alias for the corresponding SERCOM instance

Sercom2

Type alias for the corresponding SERCOM instance

Sercom3

Type alias for the corresponding SERCOM instance

Sercom4

Type alias for the corresponding SERCOM instance

Sercom5

Type alias for the corresponding SERCOM instance

Sercom6

Type alias for the corresponding SERCOM instance

Sercom7

Type alias for the corresponding SERCOM instance