atsamd_hal::sercom::i2c
Struct Status

pub struct Status { /* private fields */ }

Status flags for I2C transactions

The available status flags are BUSERR, ARBLOST, RXNACK, BUSSTATE, LOWTOUT and CLKHOLD, MEXTTOUT, SEXTTOUT and LENERR. The binary format of the underlying bits exactly matches the STATUS bits.

Implementations

impl Status

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl Status

pub const fn into_bytes(self) -> [u8; 2]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 2]) -> Self

Converts the given bytes directly into the bitfield struct.

impl Status

pub fn buserr(&self) -> <bool as Specifier>::InOut

Returns the value of buserr.

pub fn buserr_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of buserr.

#Errors

If the returned value contains an invalid bit pattern for buserr.

pub fn with_buserr(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of buserr set to the given value.

#Panics

If the given value is out of bounds for buserr.

pub fn with_buserr_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of buserr set to the given value.

#Errors

If the given value is out of bounds for buserr.

pub fn set_buserr(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of buserr to the given value.

#Panics

If the given value is out of bounds for buserr.

pub fn set_buserr_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of buserr to the given value.

#Errors

If the given value is out of bounds for buserr.

pub fn arblost(&self) -> <bool as Specifier>::InOut

Returns the value of arblost.

pub fn arblost_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of arblost.

#Errors

If the returned value contains an invalid bit pattern for arblost.

pub fn with_arblost(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of arblost set to the given value.

#Panics

If the given value is out of bounds for arblost.

pub fn with_arblost_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of arblost set to the given value.

#Errors

If the given value is out of bounds for arblost.

pub fn set_arblost(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of arblost to the given value.

#Panics

If the given value is out of bounds for arblost.

pub fn set_arblost_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of arblost to the given value.

#Errors

If the given value is out of bounds for arblost.

pub fn rxnack(&self) -> <bool as Specifier>::InOut

Returns the value of rxnack.

pub fn rxnack_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of rxnack.

#Errors

If the returned value contains an invalid bit pattern for rxnack.

pub fn busstate(&self) -> <BusState as Specifier>::InOut

Returns the value of busstate.

pub fn busstate_or_err( &self, ) -> Result<<BusState as Specifier>::InOut, InvalidBitPattern<<BusState as Specifier>::Bytes>>

Returns the value of busstate.

#Errors

If the returned value contains an invalid bit pattern for busstate.

pub fn with_busstate(self, new_val: <BusState as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of busstate set to the given value.

#Panics

If the given value is out of bounds for busstate.

pub fn with_busstate_checked( self, new_val: <BusState as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of busstate set to the given value.

#Errors

If the given value is out of bounds for busstate.

pub fn set_busstate(&mut self, new_val: <BusState as Specifier>::InOut)

Sets the value of busstate to the given value.

#Panics

If the given value is out of bounds for busstate.

pub fn set_busstate_checked( &mut self, new_val: <BusState as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of busstate to the given value.

#Errors

If the given value is out of bounds for busstate.

pub fn lowtout(&self) -> <bool as Specifier>::InOut

Returns the value of lowtout.

pub fn lowtout_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of lowtout.

#Errors

If the returned value contains an invalid bit pattern for lowtout.

pub fn with_lowtout(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of lowtout set to the given value.

#Panics

If the given value is out of bounds for lowtout.

pub fn with_lowtout_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of lowtout set to the given value.

#Errors

If the given value is out of bounds for lowtout.

pub fn set_lowtout(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of lowtout to the given value.

#Panics

If the given value is out of bounds for lowtout.

pub fn set_lowtout_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of lowtout to the given value.

#Errors

If the given value is out of bounds for lowtout.

pub fn clkhold(&self) -> <bool as Specifier>::InOut

Returns the value of clkhold.

pub fn clkhold_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of clkhold.

#Errors

If the returned value contains an invalid bit pattern for clkhold.

pub fn mexttout(&self) -> <bool as Specifier>::InOut

Returns the value of mexttout.

pub fn mexttout_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of mexttout.

#Errors

If the returned value contains an invalid bit pattern for mexttout.

pub fn with_mexttout(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of mexttout set to the given value.

#Panics

If the given value is out of bounds for mexttout.

pub fn with_mexttout_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of mexttout set to the given value.

#Errors

If the given value is out of bounds for mexttout.

pub fn set_mexttout(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of mexttout to the given value.

#Panics

If the given value is out of bounds for mexttout.

pub fn set_mexttout_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of mexttout to the given value.

#Errors

If the given value is out of bounds for mexttout.

pub fn sexttout(&self) -> <bool as Specifier>::InOut

Returns the value of sexttout.

pub fn sexttout_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of sexttout.

#Errors

If the returned value contains an invalid bit pattern for sexttout.

pub fn with_sexttout(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of sexttout set to the given value.

#Panics

If the given value is out of bounds for sexttout.

pub fn with_sexttout_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of sexttout set to the given value.

#Errors

If the given value is out of bounds for sexttout.

pub fn set_sexttout(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of sexttout to the given value.

#Panics

If the given value is out of bounds for sexttout.

pub fn set_sexttout_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of sexttout to the given value.

#Errors

If the given value is out of bounds for sexttout.

pub fn lenerr(&self) -> <bool as Specifier>::InOut

Returns the value of lenerr.

pub fn lenerr_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of lenerr.

#Errors

If the returned value contains an invalid bit pattern for lenerr.

pub fn with_lenerr(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of lenerr set to the given value.

#Panics

If the given value is out of bounds for lenerr.

pub fn with_lenerr_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of lenerr set to the given value.

#Errors

If the given value is out of bounds for lenerr.

pub fn set_lenerr(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of lenerr to the given value.

#Panics

If the given value is out of bounds for lenerr.

pub fn set_lenerr_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of lenerr to the given value.

#Errors

If the given value is out of bounds for lenerr.

impl Status

pub fn check_bus_error(self) -> Result<(), Error>

Check whether Self originates from an error.

Errors

Returns an error if STATUS contains:

BUSERR - Bus Error

ARBLOST - Arbitration lost

LENERR - Length error

RXNACK - Receive not acknowledged

pub fn is_idle(self) -> bool