atsamd_hal::sercom::pad

Trait GetPad

pub trait GetPad<S, N>
where
    S: Sercom,
    N: PadNum,
    Self: PinId,
{
    type PinMode: PinMode;
}

Type-level function mapping PinIds to SERCOM-pad-related types

For SAMD21 and SAMx5x chips, a Sercom and a PinId is enough information to uniquely identify a pad, so this trait returns the corresponding PadNum and PinMode.

For SAMD11 chips, on the other hand, some PinIds can serve as two different PadNums for the same Sercom. For these chips, GetPad requires a second type parameter to specify the PadNum and only returns the PinMode.

See the documentation on type-level functions for more details.

Required Associated Types

type PinMode: PinMode

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.