atsamd_hal::sercom::pad

Trait InIoSet

pub trait InIoSet<I>
where
    Self: IsPad,
    I: IoSet,
{ }

Type class for SERCOM pads in a given IoSet

This trait is used to label each Pin implementing IsPad with its corresponding IoSet(s). Downstream types can use this trait as a type class to restrict Pins to a given IoSet. See the type class documentation for more details on the pattern.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.