atsamd_hal::sercom::spi
Struct Spi

pub struct Spi<C, A, RxDma = NoneT, TxDma = NoneT>
where
    C: ValidConfig,
    A: Capability,
{ /* private fields */ }

An enabled SPI peripheral that can perform transactions

See the impl_ehal documentation for details on the implementations of the embedded HAL traits, which vary based on Size and Capability.

Implementations

impl<C, A, S> Spi<C, A>
where C: ValidConfig<Sercom = S>, A: Capability, S: Sercom,

pub fn into_future<I>(self, _interrupts: I) -> SpiFuture<C, A>
where C::Word: Copy, u8: AsPrimitive<C::Word>, I: Binding<S::Interrupt, InterruptHandler<S>>,

Turn an Spi into a SpiFuture.

In cases where the underlying Spi is Duplex, reading words need to be accompanied with sending a no-op word. By default it is set to 0x00, but you can configure it by using the nop_word method.

impl<C, A, RxDma, TxDma> Spi<C, A, RxDma, TxDma>
where C: ValidConfig, A: Capability,

pub fn length<L: Length>( self, ) -> Spi<Config<C::Pads, C::OpMode, L>, A, RxDma, TxDma>
where Config<C::Pads, C::OpMode, L>: ValidConfig,

Change the transaction Length

Changing the transaction Length while is enabled is permissible but dangerous. If you have sent or received any bytes at the current Length, you must wait for a TXC flag before changing to a new Length.

pub fn reconfigure(&mut self, update: impl FnOnce(&mut SpecificConfig<C>))

Update the SPI configuration.

Calling this method will temporarily disable the SERCOM peripheral, as some registers are enable-protected. This may interrupt any ongoing transactions.

pub fn enable_interrupts(&mut self, flags: Flags)

Enable interrupts for the specified flags

pub fn disable_interrupts(&mut self, flags: Flags)

Disable interrupts for the specified flags

pub fn read_flags(&self) -> Flags

Read the interrupt flags

pub fn clear_flags(&mut self, flags: Flags)

Clear the corresponding interrupt flags

Only the ERROR, SSL and TXC flags can be cleared.

Note: The implementation of of serial::Write::flush waits on and clears the TXC flag. Manually clearing this flag could cause it to hang indefinitely.

pub fn read_status(&self) -> Status

Read the error status flags

pub fn clear_status(&mut self, status: Status)

Clear the corresponding error status flags

pub fn read_flags_errors(&self) -> Result<Flags, Error>

Try to read the interrupt flags, but first check the error status flags.

pub unsafe fn read_data(&mut self) -> DataWidth

Read from the DATA register

Safety

Reading from the data register directly is unsafe, because it will clear the RXC flag, which could break assumptions made elsewhere in this module.

pub unsafe fn write_data(&mut self, data: DataWidth)

Write to the DATA register

Safety

Writing to the data register directly is unsafe, because it will clear the DRE flag, which could break assumptions made elsewhere in this module.

pub fn disable(self) -> C

Disable the SPI peripheral and return the Config struct

impl<C, D> Spi<C, D>
where C: ValidConfig, D: Receive, C::OpMode: MasterMode,

pub fn with_dma_channels<R, T>(self, rx: R, tx: T) -> Spi<C, D, R, T>
where R: AnyChannel<Status = Ready>, T: AnyChannel<Status = Ready>,

Attach RX and TX DMA channels to this Spi. Its SpiBus implementation will use DMA to carry out its transactions. In Master mode, since even read SPI transaction necessarily involve a write, Rx-only must take two DMA channels, just the same as if it were Duplex.

impl<C, D, RxDma, TxDma, S> Spi<C, D, RxDma, TxDma>
where C: ValidConfig, D: Capability, RxDma: AnyChannel<Status = S>, TxDma: AnyChannel<Status = S>, S: ReadyChannel,

pub fn take_dma_channels(self) -> (Spi<C, D, NoneT, NoneT>, RxDma, TxDma)

Reclaim the DMA channels. Any subsequent SPI transaction will not use DMA.

impl<C> Spi<C, Duplex>
where C: ValidConfig<OpMode = Slave>,

pub fn with_dma_channels_slave<R, T>(self, rx: R, tx: T) -> Spi<C, Duplex, R, T>
where R: AnyChannel<Status = Ready>, T: AnyChannel<Status = Ready>,

Attach a DMA channel to this Spi. Its SpiBus implementation will use DMA to carry out its transactions. In Slave mode, a Duplex Spi needs two DMA channels.

impl<C> Spi<C, Rx>
where C: ValidConfig<OpMode = Slave>,

pub fn with_rx_channel<R>(self, rx: R) -> Spi<C, Rx, R, NoneT>
where R: AnyChannel<Status = Ready>,

Attach a DMA channel to this Spi. Its SpiBus implementation will use DMA to carry out its transactions. In Slave mode, a Rx Spi only needs a single DMA channel.

impl<C, D, R, T, S> Spi<C, D, R, T>
where C: ValidConfig, D: Receive, R: AnyChannel<Status = S>, S: ReadyChannel,

pub fn take_rx_channel(self) -> (Spi<C, D, NoneT, T>, R)

Reclaim the Rx DMA channel. Any subsequent SPI transaction will not use DMA.

impl<C> Spi<C, Tx>
where C: ValidConfig,

pub fn with_tx_channel<T>(self, tx: T) -> Spi<C, Tx, NoneT, T>
where T: AnyChannel<Status = Ready>,

Attach a DMA channel to this Spi. Its SpiBus implementation will use DMA to carry out its transactions. For Tx Spis, only a single DMA channel is necessary.

impl<C, D, R, T, S> Spi<C, D, R, T>
where C: ValidConfig, D: Capability, T: AnyChannel<Status = S>, S: ReadyChannel,

pub fn take_tx_channel(self) -> (Spi<C, D, R, NoneT>, T)

Reclaim the DMA channel. Any subsequent SPI transaction will not use DMA.

impl<C: ValidConfig, R, T> Spi<C, Tx, R, T>

pub fn into_panic_on_write(self) -> PanicOnWrite<Self>

Turn a Tx Spi into a PanicOnWrite

impl<C: ValidConfig, R, T> Spi<C, Tx, R, T>

pub fn into_panic_on_read(self) -> PanicOnRead<Self>

Turn a Rx Spi into a PanicOnRead

impl<P, M, A> Spi<Config<P, M, DynLength>, A>
where P: ValidPads, M: OpMode, Config<P, M, DynLength>: ValidConfig, A: Capability,

pub fn get_dyn_length(&self) -> u8

Return the current transaction length

Read the LENGTH register to determine the current transaction length

pub fn set_dyn_length(&mut self, length: u8)

Set the transaction length

Write the LENGTH register to set the transaction length. Panics if the length is zero.

Changing the transaction LENGTH while is enabled is permissible but dangerous. If you have sent or received any bytes at the current LENGTH, you must wait for a TXC flag before changing to a new LENGTH.

impl<C, A> Spi<C, A>
where C: ValidConfig, A: Transmit, Self: Buffer<Beat = C::Word>,

pub fn send_with_dma<Ch, B>( self, buf: B, channel: Ch, ) -> Transfer<Channel<Ch::Id, Busy>, BufferPair<B, Self>>
where Ch: AnyChannel<Status = Ready>, B: Buffer<Beat = C::Word> + 'static,

👎Deprecated since 0.19.0: Use Spi::with_dma_channels instead. You will have access to DMA-enabled embedded-hal implementations.

Transform an Spi into a DMA Transfer) and start a send transaction.

impl<C, A> Spi<C, A>
where C: ValidConfig, A: Receive, Self: Buffer<Beat = C::Word>,

pub fn receive_with_dma<Ch, B>( self, buf: B, channel: Ch, ) -> Transfer<Channel<Ch::Id, Busy>, BufferPair<Self, B>>
where Ch: AnyChannel<Status = Ready>, B: Buffer<Beat = C::Word> + 'static,

👎Deprecated since 0.19.0: Use Spi::with_dma_channels instead. You will have access to DMA-enabled embedded-hal implementations.

Transform an Spi into a DMA Transfer) and start a receive transaction.