atsamd_hal::sercom::spi

Struct SpiFuture

pub struct SpiFuture<C, A, R = NoneT, T = NoneT>
where
    C: ValidConfig,
    A: Capability,
{ /* private fields */ }

async version of Spi.

Create this struct by calling Spi::into_future.

Implementations

impl<C, D, T> SpiFuture<C, D, NoneT, T>
where C: ValidConfig<Sercom: Sercom, OpMode = Slave, Word: PrimInt + AsPrimitive<DataWidth>>, D: Receive, DataWidth: AsPrimitive<C::Word>,

pub fn with_rx_dma_channel<Chan: AnyChannel<Status = ReadyFuture>>( self, rx_channel: Chan, ) -> SpiFuture<C, D, Chan, T>

Attach a DMA channel to this SpiFuture. Its SpiBus implementation will use DMA to carry out its transactions. In Slave mode, a Rx SpiFuture only needs a single DMA channel.

impl<C, D, R> SpiFuture<C, D, R, NoneT>
where C: ValidConfig<Sercom: Sercom, Word: PrimInt + AsPrimitive<DataWidth>>, D: Transmit, DataWidth: AsPrimitive<C::Word>,

pub fn with_tx_dma_channel<Chan: AnyChannel<Status = ReadyFuture>>( self, tx_channel: Chan, ) -> SpiFuture<C, D, R, Chan>

Attach a DMA channel to this SpiFuture. Its SpiBus implementation will use DMA to carry out its transactions. For Tx SpiFutures, only a single DMA channel is necessary.

impl<C, D> SpiFuture<C, D, NoneT, NoneT>
where C: ValidConfig<Sercom: Sercom, OpMode: MasterMode, Word: PrimInt + AsPrimitive<DataWidth>>, D: Receive, DataWidth: AsPrimitive<C::Word>,

pub fn with_dma_channels<R, T>( self, rx_channel: R, tx_channel: T, ) -> SpiFuture<C, D, R, T>
where R: AnyChannel<Status = ReadyFuture>, T: AnyChannel<Status = ReadyFuture>,

Attach RX and TX DMA channels to this SpiFuture. Its SpiBus implementation will use DMA to carry out its transactions. In Master mode, since even read SPI transaction necessarily involve a write to shift data in, Rx-only must take two DMA channels, just the same as if it were Duplex.

impl<C> SpiFuture<C, Duplex>
where C: ValidConfig<OpMode = Slave>,

pub fn with_dma_channels_slave<R, T>( self, rx: R, tx: T, ) -> SpiFuture<C, Duplex, R, T>
where R: AnyChannel<Status = ReadyFuture>, T: AnyChannel<Status = ReadyFuture>,

Attach a DMA channel to this SpiFuture. Its SpiBus implementation will use DMA to carry out its transactions. In Slave mode, a Duplex SpiFuture needs two DMA channels.

impl<C, D, R, T> SpiFuture<C, D, R, T>
where C: ValidConfig, D: Capability,

pub fn take_dma_channels(self) -> (SpiFuture<C, D, NoneT, NoneT>, R, T)
where R: AnyChannel<Status = ReadyFuture>, T: AnyChannel<Status = ReadyFuture>,

Reclaim both RX and TX DMA channels. Any subsequent SPI transaction will not use DMA.

pub fn take_rx_channel(self) -> (SpiFuture<C, D, NoneT, T>, R)
where R: AnyChannel<Status = ReadyFuture>,

Reclaim the RX DMA channel. Any subsequent SPI RX transaction will not use DMA.

pub fn take_tx_channel(self) -> (SpiFuture<C, D, R, NoneT>, T)
where T: AnyChannel<Status = ReadyFuture>,

Reclaim the TX DMA channel. Any subsequent SPI TX transaction will not use DMA.

impl<P, M, Z, D, R, T, S> SpiFuture<Config<P, M, Z>, D, R, T>
where P: ValidPads, M: OpMode, Z: Size + 'static, Config<P, M, Z>: ValidConfig<Sercom = S>, D: Transmit, S: Sercom, Z::Word: PrimInt + AsPrimitive<DataWidth> + Beat, DataWidth: AsPrimitive<Z::Word>, T: AnyChannel<Status = ReadyFuture>,

pub async fn write_dma(&mut self, words: &[Z::Word]) -> Result<usize, Error>

Write words from a buffer asynchronously, using DMA.

impl<P, M, S, C, D, R, T> SpiFuture<Config<P, M, C>, D, R, T>
where Config<P, M, C>: ValidConfig<Sercom = S>, S: Sercom, P: ValidPads, M: MasterMode, C: Size + 'static, C::Word: PrimInt + AsPrimitive<DataWidth> + Beat, D: Capability, DataWidth: AsPrimitive<C::Word>, R: AnyChannel<Status = ReadyFuture>, T: AnyChannel<Status = ReadyFuture>,

pub async fn read_dma_master( &mut self, words: &mut [C::Word], ) -> Result<(), Error>

Read words into a buffer asynchronously, using DMA.

impl<C, D> SpiFuture<C, D, NoneT, NoneT>
where C: ValidConfig, D: Capability,

pub fn free(self) -> Spi<C, D>

Return the underlying Spi.

impl<P, M, C, D, R, S> SpiFuture<Config<P, M, C>, D, R, NoneT>
where Config<P, M, C>: ValidConfig<Sercom = S>, P: ValidPads, M: OpMode, C: Size + 'static, C::Word: PrimInt + AsPrimitive<DataWidth>, D: Transmit, DataWidth: AsPrimitive<C::Word>, S: Sercom,

pub async fn write(&mut self, words: &[C::Word]) -> Result<(), Error>

Write words from a buffer asynchronously, word by word

impl<P, M, C, D, T, S> SpiFuture<Config<P, M, C>, D, NoneT, T>
where Config<P, M, C>: ValidConfig<Sercom = S>, P: ValidPads, M: MasterMode, C: Size + 'static, C::Word: PrimInt + AsPrimitive<DataWidth>, D: Receive, DataWidth: AsPrimitive<C::Word>, S: Sercom,

pub async fn read(&mut self, buffer: &mut [C::Word]) -> Result<(), Error>

Read words into a buffer asynchronously, word by word.

Since we are using a Duplex SpiFuture, we need to send a word simultaneously while receiving one. This no-op word is configurable via Config::set_nop_word.

impl<P, M, C, S, D, R, T> SpiFuture<Config<P, M, C>, D, R, T>
where Config<P, M, C>: ValidConfig<Sercom = S>, P: ValidPads, M: OpMode, C: Size + 'static, C::Word: PrimInt + AsPrimitive<DataWidth>, DataWidth: AsPrimitive<C::Word>, D: Capability, S: Sercom,

pub async fn transfer_word_in_place( &mut self, to_send: C::Word, ) -> Result<C::Word, Error>

Read and write a single word to the bus simultaneously.