atsamd_hal::sercom::spi
Trait AnySpi

pub trait AnySpi: Is<Type = SpecificSpi<Self>> {
    type Sercom: Sercom;
    type Pads: ValidPads;
    type Capability: Capability;
    type OpMode: OpMode;
    type Size: Size;
    type Word: 'static;
    type Config: ValidConfig<Sercom = Self::Sercom>;
}

Type class for all possible Spi types

This trait uses the AnyKind trait pattern to create a type class for Spi types. See the AnyKind documentation for more details on the pattern.

In addition to the normal, AnyKind associated types. This trait also copies the Sercom, Pads, Capability, OpMode, Size and Word types, to make it easier to apply bounds to these types at the next level of abstraction.

Required Associated Types

type Sercom: Sercom

type Pads: ValidPads

type Capability: Capability

type OpMode: OpMode

type Size: Size

type Word: 'static

type Config: ValidConfig<Sercom = Self::Sercom>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.