atsamd_hal::sercom::spi
Trait OpMode

pub trait OpMode: Sealed {
    const MODE: Modeselect;
    const MSSEN: bool;
}

Type-level enum representing the SPI operating mode

See the documentation on type-level enums for a discussion of the pattern.

The available operating modes are Master, MasterHWSS and Slave. In Master mode, the SS signal must be handled by the user, so SS must be NoneT. In MasterHWSS mode, the hardware drives the SS line, so SomePad is required. In Slave mode, the SS pad is required as well, to indicate when data is valid.

Required Associated Constants

const MODE: Modeselect

Corresponding variant from the PAC enum

const MSSEN: bool

Bit indicating whether hardware SS control is enabled

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.