atsamd_hal::sercom::spi_future
Trait CheckBufLen

pub trait CheckBufLen: AnySpi {
    const LEN: usize = <Self::Size as Unsigned>::USIZE;
    const STEP: usize = _;

    // Provided methods
    fn len(&self) -> usize { ... }
    fn step(&self) -> usize { ... }
    fn check_buf_len(&self, buf: &impl AsRef<[u8]>) { ... }
}

👎Deprecated since 0.19.0: spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the async APIs available in the spi module as a replacement.

Trait used to verify the SpiFuture buffer length

Provided Associated Constants

const LEN: usize = <Self::Size as Unsigned>::USIZE

👎Deprecated since 0.19.0: spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the async APIs available in the spi module as a replacement.

Spi transaction length

This value is zero for an Spi with DynLength

const STEP: usize = _

👎Deprecated since 0.19.0: spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the async APIs available in the spi module as a replacement.

Step size through the SpiFuture buffer

This is equal to the number of bytes sent in each write to the SPI DATA register. It is zero for an Spi with DynLength.

Provided Methods

fn len(&self) -> usize

👎Deprecated since 0.19.0: spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the async APIs available in the spi module as a replacement.

Return the Spi transaction length

In most cases, this returns the corresponding constant. For SAMx5x chips with DynLength, this returns the result of Spi::get_dyn_length.

fn step(&self) -> usize

👎Deprecated since 0.19.0: spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the async APIs available in the spi module as a replacement.

Return the step size through the SpiFuture buffer

In most cases, this returns the corresponding constant. For SAMx5x chips with DynLength, this returns a result calculated from Self::len.

fn check_buf_len(&self, buf: &impl AsRef<[u8]>)

👎Deprecated since 0.19.0: spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the async APIs available in the spi module as a replacement.

Check that the buffer has a valid length

If the transaction length is greater than four, then the size of the buffer must equal the transaction length. If the transaction length is less than or equal to four, then the size of the buffer must be an integer multiple of the transaction length.

Both of these statements apply regardless of whether the Spi has a DynLength.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.