atsamd_hal::sercom::spi_future
Trait ControlSS

pub trait ControlSS: OptionalPin {
    // Required methods
    fn assert(&mut self);
    fn deassert(&mut self);
}

👎Deprecated since 0.19.0: spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the async APIs available in the spi module as a replacement.

Trait used to control the SS line during an SpiFuture transaction

Required Methods

fn assert(&mut self)

👎Deprecated since 0.19.0: spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the async APIs available in the spi module as a replacement.

If an SS pin is present, assert it by bringing it low

fn deassert(&mut self)

👎Deprecated since 0.19.0: spi_future is deprecated and will be removed in a later version of atsamd_hal. Consider using the async APIs available in the spi module as a replacement.

If an SS pin is present, deassert it by bringing it high