atsamd_hal::sercom::uart

Trait ValidPads

pub trait ValidPads: PadSet + RxpoTxpo {
    type Capability: Capability;
}

Marker trait for valid sets of Pads

This trait labels sets of Pads that satisfy the RxpoTxpo trait. It guarantees to the Config struct that this set of Pads can be configured through that trait.

Required Associated Types

type Capability: Capability

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.