Trait atsamd_hal::sercom::v2::pad::GetPad

pub trait GetPad<S> where
    S: Sercom,
    Self: PinId, {
    type PadNum: PadNum;
    type PinMode: PinMode;
}

Type-level function mapping PinIds to SERCOM-pad-related types

For SAMD21 and SAMx5x chips, a Sercom and a PinId is enough information to uniquely identify a pad, so this trait returns the corresponding PadNum and PinMode.

For SAMD11 chips, on the other hand, some PinIds can serve as two different PadNums for the same Sercom. For these chips, GetPad requires a second type parameter to specify the PadNum and only returns the PinMode.

See the documentation on type-level functions for more details.

Associated Types

type PadNum: PadNum

type PinMode: PinMode

Loading content...